@@ -1,95 +1,173 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mobile_pos/Screens/Authentication/forgot_password.dart';
-import 'package:mobile_pos/Screens/Authentication/login_form.dart';
-import 'package:mobile_pos/Screens/Authentication/register_screen.dart';
-import 'package:mobile_pos/Screens/Authentication/sign_in.dart';
-import 'package:mobile_pos/Screens/Customers/customer_list.dart';
-import 'package:mobile_pos/Screens/Expense/expense_list.dart';
-import 'package:mobile_pos/Screens/Home/home.dart';
-import 'package:mobile_pos/Screens/Payment/payment_options.dart';
-import 'package:mobile_pos/Screens/Products/add_product.dart';
-import 'package:mobile_pos/Screens/Products/product_list_screen.dart';
-import 'package:mobile_pos/Screens/Profile/profile_screen.dart';
-import 'package:mobile_pos/Screens/Report/reports.dart';
-import 'package:mobile_pos/Screens/Sales/add_discount.dart';
-import 'package:mobile_pos/Screens/Sales/add_promo_code.dart';
-import 'package:mobile_pos/Screens/Sales/sales_contact.dart';
-import 'package:mobile_pos/Screens/SplashScreen/on_board.dart';
-import 'package:mobile_pos/Screens/SplashScreen/splash_screen.dart';
-import 'package:mobile_pos/Screens/stock_list/stock_list.dart';
-import 'package:provider/provider.dart' as pro;
-import 'package:shurjopay/utilities/functions.dart';
-import 'Screens/Due Calculation/due_calculation_contact_screen.dart';
-import 'Screens/Home/home_screen.dart';
-import 'Screens/Loss_Profit/loss_profit_screen.dart';
-import 'Screens/Products/update_product.dart';
-import 'Screens/Purchase List/purchase_list_screen.dart';
-import 'Screens/Purchase/choose_supplier_screen.dart';
-import 'Screens/Sales List/sales_list_screen.dart';
-import 'Screens/language/language_provider.dart';
-import 'generated/l10n.dart';
+import 'package:provider/provider.dart';
+import 'Provider/app_provider.dart';
+import 'Provider/auth_provider.dart';
+import 'Provider/ai_provider.dart';
+import 'Provider/web3_provider.dart';
+import 'Provider/blockchain_provider.dart';
+import 'Provider/analytics_provider.dart';
+import 'Provider/social_provider.dart';
+import 'Provider/product_provider.dart';
+import 'Provider/inventory_provider.dart';
+import 'Provider/nft_provider.dart';
+import 'Screens/main_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  initializeShurjopay(environment: 'live');
-  runApp(
-    const ProviderScope(
-      child: MyApp(),
-    ),
-  );
+  
+  // Создаем и инициализируем AppProvider
+  final appProvider = AppProvider();
+  await appProvider.initialize();
+  
+  runApp(PoSProApp(appProvider: appProvider));
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+class PoSProApp extends StatelessWidget {
+  final AppProvider appProvider;
+  
+  const PoSProApp({super.key, required this.appProvider});
+
   @override
   Widget build(BuildContext context) {
-    return pro.ChangeNotifierProvider<LanguageChangeProvider>(
-      create: (context) => LanguageChangeProvider(),
-      child: Builder(
-          builder: (context) => MaterialApp(
-                debugShowCheckedModeBanner: false,
-                locale: pro.Provider.of<LanguageChangeProvider>(context, listen: true).currentLocale,
-                localizationsDelegates: const [
-                  S.delegate,
-                  GlobalMaterialLocalizations.delegate,
-                  GlobalWidgetsLocalizations.delegate,
-                  GlobalCupertinoLocalizations.delegate,
-                ],
-                supportedLocales: S.delegate.supportedLocales,
-                title: 'POSpro',
-                initialRoute: '/',
-                builder: EasyLoading.init(),
-                routes: {
-                  '/': (context) => const SplashScreen(),
-                  '/onBoard': (context) => const OnBoard(),
-                  '/signIn': (context) => const SignInScreen(),
-                  '/loginForm': (context) => const LoginForm(isEmailLogin: true),
-                  '/signup': (context) => const RegisterScreen(),
-                  '/forgotPassword': (context) => const ForgotPassword(),
-                  '/home': (context) => const Home(),
-                  '/profile': (context) => const ProfileScreen(),
-                  '/AddProducts': (context) => AddProduct(),
-                  '/UpdateProducts': (context) => UpdateProduct(),
-                  '/Products': (context) => const ProductList(),
-                  '/salesCustomer': (context) => const SalesContact(),
-                  '/addPromoCode': (context) => const AddPromoCode(),
-                  '/addDiscount': (context) => const AddDiscount(),
-                  '/Sales': (context) => const SalesContact(),
-                  '/Parties': (context) => const CustomerList(),
-                  '/Expense': (context) => const ExpenseList(),
-                  '/Stock': (context) => const StockList(isFromReport: false,),
-                  '/Purchase': (context) => const PurchaseContacts(),
-                  '/Reports': (context) => const Reports(),
-                  '/Due List': (context) => const DueCalculationContactScreen(),
-                  '/PaymentOptions': (context) => const PaymentOptions(),
-                  '/Sales List': (context) => const SalesListScreen(),
-                  '/Purchase List': (context) => const PurchaseListScreen(),
-                  '/Loss/Profit': (context) => const LossProfitScreen(),
-                },
-              )),
+    return MultiProvider(
+      providers: [
+        // Основной провайдер приложения (уже инициализированный)
+        ChangeNotifierProvider.value(value: appProvider),
+        
+        // Провайдеры для различных функций (используем те же экземпляры)
+        ChangeNotifierProvider.value(value: appProvider.authProvider),
+        ChangeNotifierProvider.value(value: appProvider.aiProvider),
+        ChangeNotifierProvider.value(value: appProvider.web3Provider),
+        ChangeNotifierProvider.value(value: appProvider.blockchainProvider),
+        ChangeNotifierProvider.value(value: appProvider.analyticsProvider),
+        ChangeNotifierProvider.value(value: appProvider.socialProvider),
+        ChangeNotifierProvider.value(value: appProvider.productProvider),
+        ChangeNotifierProvider.value(value: appProvider.aiChatProvider),
+        ChangeNotifierProvider.value(value: appProvider.posProvider),
+        ChangeNotifierProvider.value(value: appProvider.inventoryProvider),
+        ChangeNotifierProvider.value(value: appProvider.nftProvider),
+      ],
+      child: MaterialApp(
+        title: 'PoSPro - AI & Web3 PoS System',
+        theme: ThemeData(
+          colorScheme: ColorScheme.fromSeed(
+            seedColor: const Color(0xFF2196F3),
+            brightness: Brightness.light,
+          ),
+          useMaterial3: true,
+          fontFamily: 'Roboto',
+          
+          // Настройки AppBar
+          appBarTheme: const AppBarTheme(
+            elevation: 0,
+            centerTitle: true,
+            backgroundColor: Color(0xFF2196F3),
+            foregroundColor: Colors.white,
+            titleTextStyle: TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.w600,
+              color: Colors.white,
+            ),
+          ),
+          
+          // Настройки кнопок
+          elevatedButtonTheme: ElevatedButtonThemeData(
+            style: ElevatedButton.styleFrom(
+              elevation: 2,
+              padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 12),
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(8),
+              ),
+            ),
+          ),
+          
+          // Настройки карточек
+          cardTheme: CardThemeData(
+            elevation: 2,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(12),
+            ),
+            margin: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
+          ),
+          
+          // Настройки полей ввода
+          inputDecorationTheme: InputDecorationTheme(
+            border: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(8),
+            ),
+            contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+          ),
+        ),
+        
+        // Темная тема
+        darkTheme: ThemeData(
+          colorScheme: ColorScheme.fromSeed(
+            seedColor: const Color(0xFF2196F3),
+            brightness: Brightness.dark,
+          ),
+          useMaterial3: true,
+          fontFamily: 'Roboto',
+          
+          // Настройки AppBar для темной темы
+          appBarTheme: const AppBarTheme(
+            elevation: 0,
+            centerTitle: true,
+            backgroundColor: Color(0xFF1976D2),
+            foregroundColor: Colors.white,
+            titleTextStyle: TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.w600,
+              color: Colors.white,
+            ),
+          ),
+          
+          // Настройки кнопок для темной темы
+          elevatedButtonTheme: ElevatedButtonThemeData(
+            style: ElevatedButton.styleFrom(
+              elevation: 2,
+              padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 12),
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(8),
+              ),
+            ),
+          ),
+          
+          // Настройки карточек для темной темы
+          cardTheme: CardThemeData(
+            elevation: 2,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(12),
+            ),
+            margin: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
+          ),
+          
+          // Настройки полей ввода для темной темы
+          inputDecorationTheme: InputDecorationTheme(
+            border: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(8),
+            ),
+            contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+          ),
+        ),
+        
+        home: const MainScreen(),
+        
+        // Настройки локализации
+        localizationsDelegates: [
+          GlobalMaterialLocalizations.delegate,
+          GlobalWidgetsLocalizations.delegate,
+          GlobalCupertinoLocalizations.delegate,
+        ],
+        supportedLocales: const [
+          Locale('ru', 'RU'),
+          Locale('en', 'US'),
+        ],
+        locale: const Locale('en', 'US'), // Временно используем английский для избежания ошибок
+        
+        // Настройки отладки
+        debugShowCheckedModeBanner: false,
+      ),
     );
   }
 }
