@@ -1,4 +1,4 @@
-import 'package:bluetooth_thermal_printer/bluetooth_thermal_printer.dart';
+// import 'package:bluetooth_thermal_printer/bluetooth_thermal_printer.dart';
 import 'package:esc_pos_utils/esc_pos_utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -12,38 +12,24 @@ final printerPurchaseProviderNotifier = ChangeNotifierProvider((ref) => PrinterP
 
 class PrinterPurchase extends ChangeNotifier {
   List availableBluetoothDevices = [];
+  bool connected = false;
   Future<void> getBluetooth() async {
-    final List? bluetooths = await BluetoothThermalPrinter.getBluetooths;
-    availableBluetoothDevices = bluetooths!;
+    // Временно отключено для сборки
+    availableBluetoothDevices = [];
     notifyListeners();
   }
 
   Future<bool> setConnect(String mac) async {
+    // Временно отключено для сборки
     bool status = false;
-    final String? result = await BluetoothThermalPrinter.connect(mac);
-    if (result == "true") {
-      connected = true;
-      status = true;
-    }
     notifyListeners();
     return status;
   }
 
   Future<bool> printPurchaseThermalInvoice({required PrintPurchaseTransactionModel printTransactionModel, required List<Details>? productList}) async {
+    // Временно отключено для сборки
     bool isPrinted = false;
-    String? isConnected = await BluetoothThermalPrinter.connectionStatus;
-    if (isConnected == "true") {
-      List<int> bytes = await getTicket(printTransactionModel: printTransactionModel, productList: productList);
-      if (productList!.isNotEmpty) {
-        await BluetoothThermalPrinter.writeBytes(bytes);
-      } else {
-        toast('No Product Found');
-      }
-
-      isPrinted = true;
-    } else {
-      isPrinted = false;
-    }
+    toast('Printing temporarily disabled');
     notifyListeners();
     return isPrinted;
   }
