@@ -1,4 +1,4 @@
-import 'package:bluetooth_thermal_printer/bluetooth_thermal_printer.dart';
+// import 'package:flutter_bluetooth_printer/flutter_bluetooth_printer.dart';
 import 'package:esc_pos_utils/esc_pos_utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
@@ -13,39 +13,24 @@ final salesPrinterProvider = ChangeNotifierProvider((ref) => SalesPrinter());
 
 class SalesPrinter extends ChangeNotifier {
   List availableBluetoothDevices = [];
+  bool connected = false;
   Future<void> getBluetooth() async {
-    final List? bluetooths = await BluetoothThermalPrinter.getBluetooths;
-    availableBluetoothDevices = bluetooths!;
+    // Временно отключено для сборки
+    availableBluetoothDevices = [];
     notifyListeners();
   }
 
   Future<bool> setConnect(String mac) async {
+    // Временно отключено для сборки
     bool status = false;
-    final String? result = await BluetoothThermalPrinter.connect(mac);
-    if (result == "true") {
-      connected = true;
-      status = true;
-    }
     notifyListeners();
     return status;
   }
 
   Future<bool> printSalesTicket({required PrintTransactionModel printTransactionModel, required List<SalesDetails>? productList}) async {
+    // Временно отключено для сборки
     bool isPrinted = false;
-    String? isConnected = await BluetoothThermalPrinter.connectionStatus;
-    if (isConnected == "true") {
-      List<int> bytes = await getSalesTicket(printTransactionModel: printTransactionModel, productList: productList);
-      if (productList!.isNotEmpty) {
-        await BluetoothThermalPrinter.writeBytes(bytes);
-        EasyLoading.showSuccess('Successfully Printed');
-      } else {
-        toast('No Product Found');
-      }
-
-      isPrinted = true;
-    } else {
-      isPrinted = false;
-    }
+    EasyLoading.showSuccess('Printing temporarily disabled');
     notifyListeners();
     return isPrinted;
   }
