@@ -9,7 +9,6 @@ import 'package:mobile_pos/GlobalComponents/button_global.dart';
 import 'package:mobile_pos/Screens/Authentication/Phone%20Auth/Repo/phone_auth_repo.dart';
 import 'package:mobile_pos/constant.dart';
 import 'package:mobile_pos/generated/l10n.dart' as lang;
-import 'package:mobile_pos/generated/l10n.dart' as lang;
 import 'package:nb_utils/nb_utils.dart';
 import 'package:pinput/pinput.dart';
 
@@ -22,12 +21,25 @@ class OTPVerify extends StatefulWidget {
   State<OTPVerify> createState() => _OTPVerifyState();
 }
 
-class _OTPVerifyState extends State<OTPVerify> {
-
+class _OTPVerifyState extends State<OTPVerify> with TickerProviderStateMixin {
   String code = '';
   FocusNode focusNode = FocusNode();
-  int _start = 60; // 2 minutes in seconds
+  int _start = 60; // 60 seconds
   late Timer _timer;
+
+  // Animation controllers
+  late AnimationController _fadeController;
+  late AnimationController _slideController;
+  late AnimationController _scaleController;
+  late AnimationController _pulseController;
+  late AnimationController _shakeController;
+
+  // Animations
+  late Animation<double> _fadeAnimation;
+  late Animation<Offset> _slideAnimation;
+  late Animation<double> _scaleAnimation;
+  late Animation<double> _pulseAnimation;
+  late Animation<double> _shakeAnimation;
 
   void _startTimer() {
     _timer = Timer.periodic(const Duration(seconds: 1), (timer) {
@@ -46,20 +58,96 @@ class _OTPVerifyState extends State<OTPVerify> {
     _startTimer();
     PhoneAuthRepo repo = PhoneAuthRepo();
     await repo.sentOTP(phoneNumber: widget.phoneNumber, context: context);
+    
+    // Add shake animation for feedback
+    _shakeController.forward().then((_) => _shakeController.reset());
   }
 
   @override
   void initState() {
     super.initState();
+    _initializeAnimations();
     _startTimer();
+  }
+
+  void _initializeAnimations() {
+    _fadeController = AnimationController(
+      duration: const Duration(milliseconds: 800),
+      vsync: this,
+    );
+    _slideController = AnimationController(
+      duration: const Duration(milliseconds: 600),
+      vsync: this,
+    );
+    _scaleController = AnimationController(
+      duration: const Duration(milliseconds: 500),
+      vsync: this,
+    );
+    _pulseController = AnimationController(
+      duration: const Duration(milliseconds: 2000),
+      vsync: this,
+    );
+    _shakeController = AnimationController(
+      duration: const Duration(milliseconds: 500),
+      vsync: this,
+    );
+
+    _fadeAnimation = Tween<double>(
+      begin: 0.0,
+      end: 1.0,
+    ).animate(CurvedAnimation(
+      parent: _fadeController,
+      curve: Curves.easeInOut,
+    ));
+
+    _slideAnimation = Tween<Offset>(
+      begin: const Offset(0, 0.3),
+      end: Offset.zero,
+    ).animate(CurvedAnimation(
+      parent: _slideController,
+      curve: Curves.easeOutCubic,
+    ));
+
+    _scaleAnimation = Tween<double>(
+      begin: 0.8,
+      end: 1.0,
+    ).animate(CurvedAnimation(
+      parent: _scaleController,
+      curve: Curves.elasticOut,
+    ));
+
+    _pulseAnimation = Tween<double>(
+      begin: 1.0,
+      end: 1.05,
+    ).animate(CurvedAnimation(
+      parent: _pulseController,
+      curve: Curves.easeInOut,
+    ));
+
+    _shakeAnimation = Tween<double>(
+      begin: 0.0,
+      end: 10.0,
+    ).animate(CurvedAnimation(
+      parent: _shakeController,
+      curve: Curves.elasticIn,
+    ));
+
+    _fadeController.forward();
+    _slideController.forward();
+    _scaleController.forward();
+    _pulseController.repeat(reverse: true);
   }
 
   @override
   void dispose() {
-    // TODO: implement dispose
-    super.dispose();
+    _fadeController.dispose();
+    _slideController.dispose();
+    _scaleController.dispose();
+    _pulseController.dispose();
+    _shakeController.dispose();
     _timer.cancel();
     focusNode.dispose();
+    super.dispose();
   }
 
   final GlobalKey<FormState> _key = GlobalKey();
@@ -73,123 +161,447 @@ class _OTPVerifyState extends State<OTPVerify> {
       child: Scaffold(
         extendBodyBehindAppBar: true,
         backgroundColor: kWhite,
-        body: Container(
-          margin: const EdgeInsets.only(left: 25, right: 25),
-          alignment: Alignment.center,
-          child: SingleChildScrollView(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                const NameWithLogo(),
-                const SizedBox(height: 25),
-                Text(
-                  lang.S.of(context).phoneVerification,
-                  style: const TextStyle(fontSize: 22, fontWeight: FontWeight.bold),
-                ),
-                const SizedBox(height: 10),
-                const Text(
-                  'We sent an OTP in your phone number',
-                  style: TextStyle(
-                    fontSize: 16,
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-                TextButton(
-                  onPressed: () {
-                    Navigator.pop(context);
-                  },
-                  child: Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      Text(
-                        widget.phoneNumber,
-                        style: const TextStyle(fontSize: 16, color: kMainColor),
-                        textAlign: TextAlign.center,
-                      ),
-                      const SizedBox(width: 10),
-                      const Icon(
-                        IconlyLight.edit_square,
-                        size: 16,
-                        color: kMainColor,
-                      )
-                    ],
+        body: CustomScrollView(
+          slivers: [
+            _buildSliverAppBar(),
+            SliverToBoxAdapter(
+              child: FadeTransition(
+                opacity: _fadeAnimation,
+                child: SlideTransition(
+                  position: _slideAnimation,
+                  child: Container(
+                    margin: const EdgeInsets.all(25),
+                    child: Column(
+                      children: [
+                        _buildLogoSection(),
+                        const SizedBox(height: 30),
+                        _buildWelcomeSection(),
+                        const SizedBox(height: 20),
+                        _buildPhoneNumberSection(),
+                        const SizedBox(height: 25),
+                        _buildOTPInputSection(),
+                        const SizedBox(height: 25),
+                        _buildActionSection(),
+                        const SizedBox(height: 20),
+                        _buildResendSection(),
+                      ],
+                    ),
                   ),
                 ),
-                const SizedBox(height: 20),
-                Form(
-                  key: _key,
-                  child: Pinput(
-                      focusNode: focusNode,
-                      keyboardType: TextInputType.number,
-                      errorPinTheme:
-                          PinTheme(width: 50, height: 50, decoration: BoxDecoration(color: Colors.red.shade200, borderRadius: const BorderRadius.all(Radius.circular(8)))),
-                      validator: (value) {
-                        if (value.isEmptyOrNull) {
-                          return 'Please enter the OTP';
-                        }
-                        if (value!.length < 4) {
-                          return 'Enter a valid OTP';
-                        } else {
-                          return null;
-                        }
-                      },
-                      length: 4,
-                      showCursor: true,
-                      onCompleted: (pin) {
-                        code = pin;
-                      }),
-                ),
-                const SizedBox(height: 20),
-                SizedBox(
-                  width: double.infinity,
-                  height: 45,
-                  child: ElevatedButton(
-                      style: ElevatedButton.styleFrom(backgroundColor: kMainColor, shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10))),
-                      onPressed: () async {
-                        focusNode.unfocus();
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
 
-                        if (_key.currentState?.validate() ?? false) {
-                          EasyLoading.show();
-
-                          PhoneAuthRepo repo = PhoneAuthRepo();
-
-                          await repo.submitOTP(phoneNumber: widget.phoneNumber, otp: code, context: context);
-                        }
-                      },
-                      child:  const Text(
-                        'Verify',
-                        style: TextStyle(color: Colors.white),
-                      )),
-                ),
-                SizedBox(
-                  height: 20,
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    _start == 0
-                        ? GestureDetector(
-                            onTap: _resendOtp,
-                            child: const Text(
-                              'Resend OTP',
-                              style: TextStyle(color: kMainColor),
-                            ))
-                        : Row(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              const Text('Resend OTP in '),
-                              Text(
-                                '${_start.toString()} seconds',
-                                style: const TextStyle(color: Colors.grey),
-                              ),
-                            ],
-                          )
-                  ],
-                )
+  Widget _buildSliverAppBar() {
+    return SliverAppBar(
+      expandedHeight: 120,
+      pinned: true,
+      backgroundColor: kMainColor,
+      flexibleSpace: FlexibleSpaceBar(
+        background: Container(
+          decoration: BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topLeft,
+              end: Alignment.bottomRight,
+              colors: [
+                kMainColor,
+                kMainColor.withOpacity(0.8),
+                kMainColor.withOpacity(0.6),
               ],
             ),
           ),
+          child: Center(
+            child: ScaleTransition(
+              scale: _pulseAnimation,
+              child: Container(
+                padding: const EdgeInsets.all(16),
+                decoration: BoxDecoration(
+                  color: Colors.white.withOpacity(0.2),
+                  borderRadius: BorderRadius.circular(20),
+                ),
+                child: const Icon(
+                  Icons.security,
+                  size: 40,
+                  color: Colors.white,
+                ),
+              ),
+            ),
+          ),
         ),
+        title: const Text(
+          'OTP Verification',
+          style: TextStyle(
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        centerTitle: true,
+      ),
+      elevation: 0,
+    );
+  }
+
+  Widget _buildLogoSection() {
+    return ScaleTransition(
+      scale: _scaleAnimation,
+      child: Container(
+        padding: const EdgeInsets.all(24),
+        decoration: BoxDecoration(
+          color: kMainColor.withOpacity(0.1),
+          borderRadius: BorderRadius.circular(20),
+          border: Border.all(
+            color: kMainColor.withOpacity(0.3),
+            width: 1,
+          ),
+        ),
+        child: Column(
+          children: [
+            Container(
+              padding: const EdgeInsets.all(16),
+              decoration: BoxDecoration(
+                color: kMainColor,
+                borderRadius: BorderRadius.circular(16),
+              ),
+              child: const Icon(
+                Icons.security,
+                size: 32,
+                color: Colors.white,
+              ),
+            ),
+            const SizedBox(height: 16),
+            const Text(
+              'POSPro',
+              style: TextStyle(
+                fontSize: 24,
+                fontWeight: FontWeight.bold,
+                color: kMainColor,
+              ),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              'Security Verification',
+              style: TextStyle(
+                fontSize: 14,
+                color: Colors.grey[600],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildWelcomeSection() {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(16),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.withOpacity(0.1),
+            spreadRadius: 1,
+            blurRadius: 10,
+            offset: const Offset(0, 2),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          const Text(
+            'Enter Verification Code',
+            style: TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.bold,
+              color: kMainColor,
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 8),
+          Text(
+            'We\'ve sent a 4-digit code to your phone number',
+            style: TextStyle(
+              fontSize: 14,
+              color: Colors.grey[600],
+            ),
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildPhoneNumberSection() {
+    return Container(
+      padding: const EdgeInsets.all(16),
+      decoration: BoxDecoration(
+        color: Colors.grey[50],
+        borderRadius: BorderRadius.circular(12),
+        border: Border.all(color: Colors.grey[200]!),
+      ),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Icon(
+            Icons.phone,
+            color: kMainColor,
+            size: 20,
+          ),
+          const SizedBox(width: 8),
+          Text(
+            widget.phoneNumber,
+            style: TextStyle(
+              fontSize: 16,
+              color: kMainColor,
+              fontWeight: FontWeight.w600,
+            ),
+          ),
+          const SizedBox(width: 12),
+          TextButton.icon(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            icon: Icon(
+              IconlyLight.edit_square,
+              size: 16,
+              color: kMainColor,
+            ),
+            label: Text(
+              'Change',
+              style: TextStyle(
+                color: kMainColor,
+                fontSize: 14,
+              ),
+            ),
+            style: TextButton.styleFrom(
+              padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
+              minimumSize: Size.zero,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildOTPInputSection() {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(16),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.withOpacity(0.1),
+            spreadRadius: 1,
+            blurRadius: 10,
+            offset: const Offset(0, 2),
+          ),
+        ],
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const Text(
+            'Verification Code',
+            style: TextStyle(
+              fontSize: 16,
+              fontWeight: FontWeight.w600,
+              color: kMainColor,
+            ),
+          ),
+          const SizedBox(height: 16),
+          Form(
+            key: _key,
+            child: Pinput(
+              focusNode: focusNode,
+              keyboardType: TextInputType.number,
+              length: 4,
+              showCursor: true,
+              onCompleted: (pin) {
+                setState(() {
+                  code = pin;
+                });
+              },
+              defaultPinTheme: PinTheme(
+                width: 60,
+                height: 60,
+                decoration: BoxDecoration(
+                  color: Colors.grey[50],
+                  borderRadius: BorderRadius.circular(12),
+                  border: Border.all(color: Colors.grey[300]!),
+                ),
+                textStyle: const TextStyle(
+                  fontSize: 20,
+                  fontWeight: FontWeight.w600,
+                  color: kMainColor,
+                ),
+              ),
+              focusedPinTheme: PinTheme(
+                width: 60,
+                height: 60,
+                decoration: BoxDecoration(
+                  color: Colors.white,
+                  borderRadius: BorderRadius.circular(12),
+                  border: Border.all(color: kMainColor, width: 2),
+                  boxShadow: [
+                    BoxShadow(
+                      color: kMainColor.withOpacity(0.2),
+                      spreadRadius: 1,
+                      blurRadius: 8,
+                      offset: const Offset(0, 2),
+                    ),
+                  ],
+                ),
+                textStyle: const TextStyle(
+                  fontSize: 20,
+                  fontWeight: FontWeight.w600,
+                  color: kMainColor,
+                ),
+              ),
+              errorPinTheme: PinTheme(
+                width: 60,
+                height: 60,
+                decoration: BoxDecoration(
+                  color: Colors.red[50],
+                  borderRadius: BorderRadius.circular(12),
+                  border: Border.all(color: Colors.red[300]!),
+                ),
+                textStyle: const TextStyle(
+                  fontSize: 20,
+                  fontWeight: FontWeight.w600,
+                  color: Colors.red,
+                ),
+              ),
+              validator: (value) {
+                if (value == null || value.isEmpty) {
+                  return 'Please enter the verification code';
+                }
+                if (value.length < 4) {
+                  return 'Please enter a complete 4-digit code';
+                }
+                return null;
+              },
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildActionSection() {
+    return SizedBox(
+      width: double.infinity,
+      height: 50,
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(
+          backgroundColor: kMainColor,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(12),
+          ),
+          elevation: 2,
+        ),
+        onPressed: code.length == 4
+            ? () async {
+                focusNode.unfocus();
+                
+                if (_key.currentState?.validate() ?? false) {
+                  EasyLoading.show();
+                  
+                  PhoneAuthRepo repo = PhoneAuthRepo();
+                  await repo.submitOTP(
+                    phoneNumber: widget.phoneNumber,
+                    otp: code,
+                    context: context,
+                  );
+                }
+              }
+            : null,
+        child: const Text(
+          'Verify & Continue',
+          style: TextStyle(
+            color: Colors.white,
+            fontSize: 16,
+            fontWeight: FontWeight.w600,
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildResendSection() {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: Colors.grey[50],
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(color: Colors.grey[200]!),
+      ),
+      child: Column(
+        children: [
+          Text(
+            'Didn\'t receive the code?',
+            style: TextStyle(
+              color: Colors.grey[600],
+              fontSize: 14,
+            ),
+          ),
+          const SizedBox(height: 12),
+          _start == 0
+              ? AnimatedBuilder(
+                  animation: _shakeAnimation,
+                  builder: (context, child) {
+                    return Transform.translate(
+                      offset: Offset(_shakeAnimation.value, 0),
+                      child: TextButton.icon(
+                        onPressed: _resendOtp,
+                        icon: Icon(
+                          Icons.refresh,
+                          color: kMainColor,
+                          size: 18,
+                        ),
+                        label: Text(
+                          'Resend Code',
+                          style: TextStyle(
+                            color: kMainColor,
+                            fontSize: 16,
+                            fontWeight: FontWeight.w600,
+                          ),
+                        ),
+                      ),
+                    );
+                  },
+                )
+              : Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Icon(
+                      Icons.timer,
+                      color: Colors.grey[500],
+                      size: 18,
+                    ),
+                    const SizedBox(width: 8),
+                    Text(
+                      'Resend code in ',
+                      style: TextStyle(
+                        color: Colors.grey[600],
+                        fontSize: 14,
+                      ),
+                    ),
+                    Text(
+                      '${_start}s',
+                      style: TextStyle(
+                        color: kMainColor,
+                        fontSize: 16,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ],
+                ),
+        ],
       ),
     );
   }
