@@ -20,21 +20,93 @@ class PhoneAuth extends StatefulWidget {
   State<PhoneAuth> createState() => _PhoneAuthState();
 }
 
-class _PhoneAuthState extends State<PhoneAuth> {
+class _PhoneAuthState extends State<PhoneAuth> with TickerProviderStateMixin {
   String? phoneNumber;
-
   bool phoneFieldValid = true;
   late StreamSubscription subscription;
   bool isDeviceConnected = false;
   bool isAlertSet = false;
 
+  // Animation controllers
+  late AnimationController _fadeController;
+  late AnimationController _slideController;
+  late AnimationController _scaleController;
+  late AnimationController _pulseController;
+
+  // Animations
+  late Animation<double> _fadeAnimation;
+  late Animation<Offset> _slideAnimation;
+  late Animation<double> _scaleAnimation;
+  late Animation<double> _pulseAnimation;
+
   @override
   void initState() {
     super.initState();
+    _initializeAnimations();
+  }
+
+  void _initializeAnimations() {
+    _fadeController = AnimationController(
+      duration: const Duration(milliseconds: 800),
+      vsync: this,
+    );
+    _slideController = AnimationController(
+      duration: const Duration(milliseconds: 600),
+      vsync: this,
+    );
+    _scaleController = AnimationController(
+      duration: const Duration(milliseconds: 500),
+      vsync: this,
+    );
+    _pulseController = AnimationController(
+      duration: const Duration(milliseconds: 2000),
+      vsync: this,
+    );
+
+    _fadeAnimation = Tween<double>(
+      begin: 0.0,
+      end: 1.0,
+    ).animate(CurvedAnimation(
+      parent: _fadeController,
+      curve: Curves.easeInOut,
+    ));
+
+    _slideAnimation = Tween<Offset>(
+      begin: const Offset(0, 0.3),
+      end: Offset.zero,
+    ).animate(CurvedAnimation(
+      parent: _slideController,
+      curve: Curves.easeOutCubic,
+    ));
+
+    _scaleAnimation = Tween<double>(
+      begin: 0.8,
+      end: 1.0,
+    ).animate(CurvedAnimation(
+      parent: _scaleController,
+      curve: Curves.elasticOut,
+    ));
+
+    _pulseAnimation = Tween<double>(
+      begin: 1.0,
+      end: 1.1,
+    ).animate(CurvedAnimation(
+      parent: _pulseController,
+      curve: Curves.easeInOut,
+    ));
+
+    _fadeController.forward();
+    _slideController.forward();
+    _scaleController.forward();
+    _pulseController.repeat(reverse: true);
   }
 
   @override
   void dispose() {
+    _fadeController.dispose();
+    _slideController.dispose();
+    _scaleController.dispose();
+    _pulseController.dispose();
     super.dispose();
   }
 
@@ -42,167 +114,345 @@ class _PhoneAuthState extends State<PhoneAuth> {
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: kWhite,
-      body: Container(
-        margin: const EdgeInsets.only(left: 25, right: 25),
-        alignment: Alignment.center,
-        child: SingleChildScrollView(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            children: [
-              const NameWithLogo(),
-              const SizedBox(height: 25),
-              Text(
-                lang.S.of(context).phoneVerification,
-                style: const TextStyle(fontSize: 22, fontWeight: FontWeight.bold),
-              ),
-              const SizedBox(height: 10),
-              Text(
-                lang.S.of(context).registerTitle,
-                style: const TextStyle(fontSize: 16),
-                textAlign: TextAlign.center,
-              ),
-              const SizedBox(height: 30),
-              IntlPhoneField(
-                decoration: const InputDecoration(
-                  labelText: 'Phone Number',
-                  border: OutlineInputBorder(borderSide: BorderSide(), borderRadius: BorderRadius.all(Radius.circular(15))),
+      body: CustomScrollView(
+        slivers: [
+          _buildSliverAppBar(),
+          SliverToBoxAdapter(
+            child: FadeTransition(
+              opacity: _fadeAnimation,
+              child: SlideTransition(
+                position: _slideAnimation,
+                child: Container(
+                  margin: const EdgeInsets.all(25),
+                  child: Column(
+                    children: [
+                      _buildLogoSection(),
+                      const SizedBox(height: 30),
+                      _buildWelcomeSection(),
+                      const SizedBox(height: 30),
+                      _buildPhoneInputSection(),
+                      const SizedBox(height: 25),
+                      _buildActionSection(),
+                      const SizedBox(height: 20),
+                      _buildAlternativeOptions(),
+                    ],
+                  ),
                 ),
-                initialCountryCode: 'BD',
-                onChanged: (phone) {
-                  phoneNumber = phone.completeNumber;
-                },
-                inputFormatters: [FilteringTextInputFormatter.allow(RegExp(r'^\d*\.?\d{0,2}'))],
               ),
-              const SizedBox(height: 20),
-              // Container(
-              //   height: 55,
-              //   decoration:
-              //       BoxDecoration(border: Border.all(width: phoneFieldValid ? 1 : 2, color: phoneFieldValid ? Colors.grey : Colors.red), borderRadius: BorderRadius.circular(10)),
-              //   child: Row(
-              //     mainAxisAlignment: MainAxisAlignment.center,
-              //     children: [
-              //       const SizedBox(width: 10),
-              //       SizedBox(
-              //         width: 40,
-              //         child: TextField(
-              //           controller: countryController,
-              //           keyboardType: TextInputType.number,
-              //           decoration: const InputDecoration(
-              //             border: InputBorder.none,
-              //           ),
-              //         ),
-              //       ),
-              //       const Text(
-              //         "|",
-              //         style: TextStyle(fontSize: 33, color: Colors.grey),
-              //       ),
-              //       const SizedBox(width: 10),
-              //       Expanded(
-              //           child: Form(
-              //         key: _key,
-              //         child: TextFormField(
-              //           controller: phoneNumberController,
-              //           inputFormatters: [FilteringTextInputFormatter.allow(RegExp(r'^\d*\.?\d'))],
-              //           validator: (value) {
-              //             if (value.isEmptyOrNull) {
-              //               setState(() {
-              //                 phoneFieldValid = false;
-              //               });
-              //               return null;
-              //             }
-              //             if (value!.length < 8) {
-              //               setState(() {
-              //                 phoneFieldValid = false;
-              //               });
-              //               return null;
-              //             } else {
-              //               setState(() {
-              //                 phoneFieldValid = true;
-              //               });
-              //
-              //               return null;
-              //             }
-              //           },
-              //           keyboardType: TextInputType.phone,
-              //           decoration: const InputDecoration(
-              //             border: InputBorder.none,
-              //             hintText: "Phone Number",
-              //           ),
-              //         ),
-              //       ))
-              //     ],
-              //   ),
-              // ),
-              // Visibility(
-              //     visible: !phoneFieldValid,
-              //     child: const Padding(
-              //       padding: EdgeInsets.only(top: 4, left: 2),
-              //       child: Text(
-              //         'Enter a valid phone number',
-              //         style: TextStyle(color: Colors.red),
-              //       ),
-              //     )),
-              // const SizedBox(height: 20),
-              SizedBox(
-                width: double.infinity,
-                height: 45,
-                child: ElevatedButton(
-                    style: ElevatedButton.styleFrom(backgroundColor: kMainColor, shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10))),
-                    // onPressed: () async {
-                    //   // const OTPVerify().launch(context);
-                    //   _key.currentState?.validate();
-                    //
-                    //   if (phoneFieldValid) {
-                    //     EasyLoading.show();
-                    //     PhoneAuthRepo repo = PhoneAuthRepo();
-                    //
-                    //     if (await repo.sentOTP(phoneNumber: countryController.text + phoneNumberController.text, context: context)) {
-                    //       OTPVerify(phoneNumber: countryController.text + phoneNumberController.text).launch(context);
-                    //     }
-                    //   }
-                    // },
-                    onPressed: () async {
-                      if ((phoneNumber?.length ?? 0) > 8) {
-                        EasyLoading.show();
-                        PhoneAuthRepo repo = PhoneAuthRepo();
+            ),
+          ),
+        ],
+      ),
+    );
+  }
 
-                        if (await repo.sentOTP(phoneNumber: phoneNumber!, context: context)) {
-                          OTPVerify(phoneNumber: phoneNumber!).launch(context);
-                        }
-                      } else {
-                        EasyLoading.showError('Enter a valid Phone Number');
-                      }
-                    },
-                    child: Text(
-                      lang.S.of(context).sendCode,
-                      style: const TextStyle(color: Colors.white),
-                    )),
+  Widget _buildSliverAppBar() {
+    return SliverAppBar(
+      expandedHeight: 120,
+      pinned: true,
+      backgroundColor: kMainColor,
+      flexibleSpace: FlexibleSpaceBar(
+        background: Container(
+          decoration: BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topLeft,
+              end: Alignment.bottomRight,
+              colors: [
+                kMainColor,
+                kMainColor.withOpacity(0.8),
+                kMainColor.withOpacity(0.6),
+              ],
+            ),
+          ),
+          child: Center(
+            child: ScaleTransition(
+              scale: _pulseAnimation,
+              child: Container(
+                padding: const EdgeInsets.all(16),
+                decoration: BoxDecoration(
+                  color: Colors.white.withOpacity(0.2),
+                  borderRadius: BorderRadius.circular(20),
+                ),
+                child: const Icon(
+                  Icons.phone_android,
+                  size: 40,
+                  color: Colors.white,
+                ),
               ),
-              const SizedBox(height: 10),
-              // Row(
-              //   mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              //   children: [
-              //     TextButton(
-              //       onPressed: () {
-              //         const LoginForm(isEmailLogin: false).launch(context);
-              //       },
-              //       child: Text(lang.S.of(context).staffLogin),
-              //     ),
-              //     Flexible(
-              //       child: TextButton(
-              //         onPressed: () {
-              //           const LoginForm(isEmailLogin: true).launch(context);
-              //         },
-              //         child: Text(
-              //           lang.S.of(context).logInWithMail,
-              //           overflow: TextOverflow.ellipsis,
-              //           maxLines: 1,
-              //         ),
-              //       ),
-              //     ),
-              //   ],
-              // )
+            ),
+          ),
+        ),
+        title: const Text(
+          'Phone Verification',
+          style: TextStyle(
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        centerTitle: true,
+      ),
+      elevation: 0,
+    );
+  }
+
+  Widget _buildLogoSection() {
+    return ScaleTransition(
+      scale: _scaleAnimation,
+      child: Container(
+        padding: const EdgeInsets.all(24),
+        decoration: BoxDecoration(
+          color: kMainColor.withOpacity(0.1),
+          borderRadius: BorderRadius.circular(20),
+          border: Border.all(
+            color: kMainColor.withOpacity(0.3),
+            width: 1,
+          ),
+        ),
+        child: Column(
+          children: [
+            Container(
+              padding: const EdgeInsets.all(16),
+              decoration: BoxDecoration(
+                color: kMainColor,
+                borderRadius: BorderRadius.circular(16),
+              ),
+              child: const Icon(
+                Icons.phone_android,
+                size: 32,
+                color: Colors.white,
+              ),
+            ),
+            const SizedBox(height: 16),
+            const Text(
+              'POSPro',
+              style: TextStyle(
+                fontSize: 24,
+                fontWeight: FontWeight.bold,
+                color: kMainColor,
+              ),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              'Phone Authentication',
+              style: TextStyle(
+                fontSize: 14,
+                color: Colors.grey[600],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildWelcomeSection() {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(16),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.withOpacity(0.1),
+            spreadRadius: 1,
+            blurRadius: 10,
+            offset: const Offset(0, 2),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          const Text(
+            'Verify Your Phone Number',
+            style: TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.bold,
+              color: kMainColor,
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 8),
+          Text(
+            'We\'ll send you a verification code to authenticate your account',
+            style: TextStyle(
+              fontSize: 14,
+              color: Colors.grey[600],
+            ),
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildPhoneInputSection() {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(16),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.withOpacity(0.1),
+            spreadRadius: 1,
+            blurRadius: 10,
+            offset: const Offset(0, 2),
+          ),
+        ],
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const Text(
+            'Phone Number',
+            style: TextStyle(
+              fontSize: 16,
+              fontWeight: FontWeight.w600,
+              color: kMainColor,
+            ),
+          ),
+          const SizedBox(height: 12),
+          IntlPhoneField(
+            decoration: InputDecoration(
+              labelText: 'Enter your phone number',
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(12),
+                borderSide: BorderSide(color: Colors.grey[300]!),
+              ),
+              enabledBorder: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(12),
+                borderSide: BorderSide(color: Colors.grey[300]!),
+              ),
+              focusedBorder: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(12),
+                borderSide: const BorderSide(color: kMainColor, width: 2),
+              ),
+              filled: true,
+              fillColor: Colors.grey[50],
+              contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+            ),
+            initialCountryCode: 'RU',
+            onChanged: (phone) {
+              setState(() {
+                phoneNumber = phone.completeNumber;
+                phoneFieldValid = (phone.completeNumber.length ?? 0) > 8;
+              });
+            },
+            inputFormatters: [FilteringTextInputFormatter.allow(RegExp(r'^\d*\.?\d{0,2}'))],
+          ),
+          if (!phoneFieldValid && phoneNumber != null)
+            Padding(
+              padding: const EdgeInsets.only(top: 8, left: 4),
+              child: Text(
+                'Please enter a valid phone number',
+                style: TextStyle(
+                  color: Colors.red[600],
+                  fontSize: 12,
+                ),
+              ),
+            ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildActionSection() {
+    return SizedBox(
+      width: double.infinity,
+      height: 50,
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(
+          backgroundColor: kMainColor,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(12),
+          ),
+          elevation: 2,
+        ),
+        onPressed: phoneFieldValid && phoneNumber != null
+            ? () async {
+                EasyLoading.show();
+                PhoneAuthRepo repo = PhoneAuthRepo();
+                
+                if (await repo.sentOTP(phoneNumber: phoneNumber!, context: context)) {
+                  OTPVerify(phoneNumber: phoneNumber!).launch(context);
+                }
+              }
+            : null,
+        child: const Text(
+          'Send Verification Code',
+          style: TextStyle(
+            color: Colors.white,
+            fontSize: 16,
+            fontWeight: FontWeight.w600,
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildAlternativeOptions() {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: Colors.grey[50],
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(color: Colors.grey[200]!),
+      ),
+      child: Column(
+        children: [
+          Text(
+            'Or continue with',
+            style: TextStyle(
+              color: Colors.grey[600],
+              fontSize: 14,
+            ),
+          ),
+          const SizedBox(height: 16),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+            children: [
+              _buildAlternativeButton(
+                icon: Icons.email,
+                label: 'Email',
+                onTap: () {
+                  Navigator.pop(context);
+                },
+              ),
+              _buildAlternativeButton(
+                icon: Icons.person,
+                label: 'Username',
+                onTap: () {
+                  Navigator.pop(context);
+                },
+              ),
             ],
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildAlternativeButton({
+    required IconData icon,
+    required String label,
+    required VoidCallback onTap,
+  }) {
+    return Expanded(
+      child: Container(
+        margin: const EdgeInsets.symmetric(horizontal: 8),
+        child: OutlinedButton.icon(
+          onPressed: onTap,
+          icon: Icon(icon, color: kMainColor),
+          label: Text(
+            label,
+            style: const TextStyle(color: kMainColor),
+          ),
+          style: OutlinedButton.styleFrom(
+            padding: const EdgeInsets.symmetric(vertical: 12),
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(12),
+            ),
+            side: BorderSide(color: kMainColor.withOpacity(0.5)),
           ),
         ),
       ),
