@@ -16,7 +16,7 @@ class SignUpScreen extends StatefulWidget {
   State<SignUpScreen> createState() => _SignUpScreenState();
 }
 
-class _SignUpScreenState extends State<SignUpScreen> {
+class _SignUpScreenState extends State<SignUpScreen> with TickerProviderStateMixin {
   ///__________Variables________________________________
   bool showPassword = true;
   bool isClicked = false;
@@ -29,197 +29,651 @@ class _SignUpScreenState extends State<SignUpScreen> {
   TextEditingController passwordTextController = TextEditingController();
   TextEditingController emailTextController = TextEditingController();
 
+  // Анимации
+  late AnimationController _fadeController;
+  late AnimationController _slideController;
+  late AnimationController _scaleController;
+  late AnimationController _pulseController;
+  late Animation<double> _fadeAnimation;
+  late Animation<Offset> _slideAnimation;
+  late Animation<double> _scaleAnimation;
+  late Animation<double> _pulseAnimation;
+
   ///________Dispose____________________________________
   @override
   void dispose() {
-    super.dispose();
+    _fadeController.dispose();
+    _slideController.dispose();
+    _scaleController.dispose();
+    _pulseController.dispose();
     nameTextController.dispose();
     passwordTextController.dispose();
     emailTextController.dispose();
+    super.dispose();
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    
+    // Инициализация анимаций
+    _fadeController = AnimationController(
+      duration: const Duration(milliseconds: 1000),
+      vsync: this,
+    );
+    _slideController = AnimationController(
+      duration: const Duration(milliseconds: 800),
+      vsync: this,
+    );
+    _scaleController = AnimationController(
+      duration: const Duration(milliseconds: 600),
+      vsync: this,
+    );
+    _pulseController = AnimationController(
+      duration: const Duration(milliseconds: 2000),
+      vsync: this,
+    );
+
+    _fadeAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
+      CurvedAnimation(parent: _fadeController, curve: Curves.easeInOut),
+    );
+    _slideAnimation = Tween<Offset>(
+      begin: const Offset(0, 0.5),
+      end: Offset.zero,
+    ).animate(CurvedAnimation(parent: _slideController, curve: Curves.easeOut));
+    _scaleAnimation = Tween<double>(begin: 0.8, end: 1.0).animate(
+      CurvedAnimation(parent: _scaleController, curve: Curves.elasticOut),
+    );
+    _pulseAnimation = Tween<double>(begin: 1.0, end: 1.05).animate(
+      CurvedAnimation(parent: _pulseController, curve: Curves.easeInOut),
+    );
+
+    _fadeController.forward();
+    _slideController.forward();
+    _scaleController.forward();
+    _pulseController.repeat(reverse: true);
   }
 
   @override
   Widget build(BuildContext context) {
     TextTheme textTheme = Theme.of(context).textTheme;
     return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: kWhite,
-        titleSpacing: 16,
-        centerTitle: true,
-        surfaceTintColor: kWhite,
-        title: Text(
-          lang.S.of(context).signIn,
-          //'Sign Up',
-          style: textTheme.titleMedium?.copyWith(fontSize: 20),
-        ),
+      backgroundColor: const Color(0xFFF8F9FA),
+      body: CustomScrollView(
+        slivers: [
+          _buildSliverAppBar(textTheme),
+          SliverToBoxAdapter(
+            child: FadeTransition(
+              opacity: _fadeAnimation,
+              child: SlideTransition(
+                position: _slideAnimation,
+                child: Padding(
+                  padding: const EdgeInsets.fromLTRB(24.0, 20.0, 24.0, 0.0),
+                  child: Form(
+                    key: key,
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        const SizedBox(height: 32),
+                        _buildLogoSection(),
+                        const SizedBox(height: 32),
+                        _buildWelcomeSection(textTheme),
+                        const SizedBox(height: 32),
+                        _buildFormSection(textTheme),
+                        const SizedBox(height: 32),
+                        _buildActionSection(textTheme),
+                        const SizedBox(height: 32),
+                        _buildSignInSection(textTheme),
+                        const SizedBox(height: 40),
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ],
       ),
-      body: SingleChildScrollView(
-        child: Padding(
-          padding: const EdgeInsets.fromLTRB(16.0, 20.0, 16.0, 0.0),
-          child: Form(
-            key: key,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: [
-                const SizedBox(height: 24,),
-                const NameWithLogo(),
-                const SizedBox(height: 24,),
-                Text(
-                  lang.S.of(context).createAFreeAccount,
-                  //'Create A Free Account',
-                  style: textTheme.titleMedium?.copyWith(fontSize: 24.0,fontWeight: FontWeight.w600),
-                ),
-                Text(
-                  lang.S.of(context).pleaseEnterYourDetails,
-                  //'Please enter your details',
-                  style: textTheme.bodyMedium?.copyWith(color: kGreyTextColor,fontSize: 16),
-                ),
-                const SizedBox(height: 24.0),
+    );
+  }
 
-                ///____________Name______________________________________________
-                TextFormField(
-                  controller: nameTextController,
-                  keyboardType: TextInputType.name,
-                  decoration: kInputDecoration.copyWith(
-                    //labelText: 'Full Name',
-                    labelText: lang.S.of(context).fullName,
-                    //hintText: 'Enter your full name',
-                    hintText: lang.S.of(context).enterYourFullName,
-                  ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      //return 'name can\'n be empty';
-                      return lang.S.of(context).nameCanNotBeEmpty;
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 20.0),
-
-                ///__________Email______________________________________________
-                TextFormField(
-                  controller: emailTextController,
-                  keyboardType: TextInputType.emailAddress,
-                  decoration: kInputDecoration.copyWith(
-                    // border: OutlineInputBorder(),
-                    // labelText: 'email',
-                    labelText: lang.S.of(context).lableEmail,
-                    //hintText: 'Enter email address',
-                    hintText: lang.S.of(context).hintEmail,
-                  ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      //return 'Email can\'n be empty';
-                      return lang.S.of(context).emailCannotBeEmpty;
-                    } else if (!value.contains('@')) {
-                      //return 'Please enter a valid email';
-                      return lang.S.of(context).pleaseEnterAValidEmail;
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 20.0),
-
-                ///___________Password_____________________________________________
-                TextFormField(
-                  controller: passwordTextController,
-                  keyboardType: TextInputType.text,
-                  obscureText: showPassword,
-                  decoration: kInputDecoration.copyWith(
-                    //labelText: 'Password',
-                    labelText: lang.S.of(context).lablePassword,
-                   // hintText: 'Enter password',
-                    hintText: lang.S.of(context).hintPassword,
-                    suffixIcon: IconButton(
-                      onPressed: () {
-                        setState(() {
-                          showPassword = !showPassword;
-                        });
-                      },
-                      icon: Icon(
-                        showPassword ? FeatherIcons.eyeOff : FeatherIcons.eye,
-                        color: kGreyTextColor,
-                      ),
-                    ),
-                  ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      //return 'Password can\'t be empty';
-                      return lang.S.of(context).passwordCannotBeEmpty;
-                    } else if (value.length < 6) {
-                      //return 'Please enter a bigger password';
-                      return lang.S.of(context).pleaseEnterABiggerPassword;
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 24.0),
-
-                ///________Button___________________________________________________
-                UpdateButton(
-                  onpressed: () async {
-                    if (isClicked) {
-                      return;
-                    }
-                    if (key.currentState?.validate() ?? false) {
-                      isClicked = true;
-                      EasyLoading.show();
-                      SignUpRepo repo = SignUpRepo();
-                      if (await repo.signUp(name: nameTextController.text, email: emailTextController.text, password: passwordTextController.text, context: context)) {
-                        if (await checkEmailForCodePupUp(email: emailTextController.text, context: context, textTheme: textTheme)) {
-                          Navigator.pushReplacement(
-                            context,
-                            MaterialPageRoute(
-                              builder: (context) => VerifyEmail(
-                                email: emailTextController.text,
-                                isFormForgotPass: false,
-                              ),
-                            ),
-                          );
-                        }
-                      } else {
-                        isClicked = false;
-                      }
-                    }
-                  },
-                  text: lang.S.of(context).signIn,
-                  //'Sign Up',
-                ),
-                const SizedBox(height: 20,),
-                Column(
-                  mainAxisSize: MainAxisSize.min,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    InkWell(
-                      highlightColor: kMainColor.withOpacity(0.1),
-                      borderRadius: BorderRadius.circular(3.0),
-                      onTap: () => Navigator.pop(context),
-                      hoverColor: kMainColor.withOpacity(0.1),
-                      child: RichText(
-                        text: TextSpan(
-                          text:
-                          lang.S.of(context).alreadyHaveAnAccount,
-                          //'Already have an account? ',
-                          style: textTheme.bodyMedium?.copyWith(color: kGreyTextColor),
-                          children: [
-                            TextSpan(
-                              text:lang.S.of(context).signIn,
-                              //'Sign In',
-                              style: textTheme.bodyMedium?.copyWith(color: kMainColor, fontWeight: FontWeight.bold),
-                            )
-                          ],
-                        ),
-                      ),
-                    ),
-                  ],
-                )
+  Widget _buildSliverAppBar(TextTheme textTheme) {
+    return SliverAppBar(
+      expandedHeight: 120,
+      floating: false,
+      pinned: true,
+      backgroundColor: kMainColor,
+      elevation: 0,
+      automaticallyImplyLeading: false,
+      flexibleSpace: FlexibleSpaceBar(
+        title: Text(
+          'Регистрация',
+          style: textTheme.titleSmall?.copyWith(
+            fontSize: 22,
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        background: Container(
+          decoration: BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topLeft,
+              end: Alignment.bottomRight,
+              colors: [
+                kMainColor,
+                kMainColor.withOpacity(0.8),
+                kMainColor.withOpacity(0.6),
               ],
+            ),
+          ),
+          child: Center(
+            child: ScaleTransition(
+              scale: _pulseAnimation,
+              child: Icon(
+                Icons.person_add_outlined,
+                size: 60,
+                color: Colors.white.withOpacity(0.3),
+              ),
             ),
           ),
         ),
       ),
     );
+  }
+
+  Widget _buildLogoSection() {
+    return ScaleTransition(
+      scale: _scaleAnimation,
+      child: Container(
+        padding: const EdgeInsets.all(24),
+        decoration: BoxDecoration(
+          color: Colors.white,
+          borderRadius: BorderRadius.circular(20),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withOpacity(0.1),
+              blurRadius: 20,
+              offset: const Offset(0, 10),
+            ),
+          ],
+        ),
+        child: Column(
+          children: [
+            Container(
+              width: 80,
+              height: 80,
+              decoration: BoxDecoration(
+                color: kMainColor.withOpacity(0.1),
+                borderRadius: BorderRadius.circular(20),
+              ),
+              child: Icon(
+                Icons.shopping_cart,
+                size: 40,
+                color: kMainColor,
+              ),
+            ),
+            const SizedBox(height: 16),
+            Text(
+              'POSPro',
+              style: TextStyle(
+                fontSize: 28,
+                fontWeight: FontWeight.bold,
+                color: kMainColor,
+              ),
+            ),
+            Text(
+              'Professional POS System',
+              style: TextStyle(
+                fontSize: 14,
+                color: Colors.grey[600],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildWelcomeSection(TextTheme textTheme) {
+    return Container(
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 5),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          Text(
+            lang.S.of(context).createAFreeAccount,
+            style: textTheme.titleMedium?.copyWith(
+              fontSize: 28.0,
+              fontWeight: FontWeight.bold,
+              color: Colors.grey[800],
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 12),
+          Text(
+            lang.S.of(context).pleaseEnterYourDetails,
+            style: textTheme.bodyMedium?.copyWith(
+              color: Colors.grey[600],
+              fontSize: 16,
+            ),
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildFormSection(TextTheme textTheme) {
+    return Container(
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 5),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          _buildNameField(textTheme),
+          const SizedBox(height: 20),
+          _buildEmailField(textTheme),
+          const SizedBox(height: 20),
+          _buildPasswordField(textTheme),
+          const SizedBox(height: 16),
+          _buildTermsSection(textTheme),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildNameField(TextTheme textTheme) {
+    return TextFormField(
+      controller: nameTextController,
+      keyboardType: TextInputType.name,
+      decoration: InputDecoration(
+        labelText: lang.S.of(context).fullName,
+        hintText: lang.S.of(context).enterYourFullName,
+        prefixIcon: Icon(Icons.person_outline, color: kMainColor),
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: Colors.grey[300]!),
+        ),
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: Colors.grey[300]!),
+        ),
+        focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: kMainColor, width: 2),
+        ),
+        filled: true,
+        fillColor: Colors.grey[50],
+        contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
+      ),
+      validator: (value) {
+        if (value == null || value.isEmpty) {
+          return lang.S.of(context).nameCanNotBeEmpty;
+        }
+        return null;
+      },
+    );
+  }
+
+  Widget _buildEmailField(TextTheme textTheme) {
+    return TextFormField(
+      controller: emailTextController,
+      keyboardType: TextInputType.emailAddress,
+      decoration: InputDecoration(
+        labelText: lang.S.of(context).lableEmail,
+        hintText: lang.S.of(context).hintEmail,
+        prefixIcon: Icon(Icons.email_outlined, color: kMainColor),
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: Colors.grey[300]!),
+        ),
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: Colors.grey[300]!),
+        ),
+        focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: kMainColor, width: 2),
+        ),
+        filled: true,
+        fillColor: Colors.grey[50],
+        contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
+      ),
+      validator: (value) {
+        if (value == null || value.isEmpty) {
+          return lang.S.of(context).emailCannotBeEmpty;
+        } else if (!value.contains('@')) {
+          return lang.S.of(context).pleaseEnterAValidEmail;
+        }
+        return null;
+      },
+    );
+  }
+
+  Widget _buildPasswordField(TextTheme textTheme) {
+    return TextFormField(
+      controller: passwordTextController,
+      keyboardType: TextInputType.text,
+      obscureText: showPassword,
+      decoration: InputDecoration(
+        labelText: lang.S.of(context).lablePassword,
+        hintText: lang.S.of(context).hintPassword,
+        prefixIcon: Icon(Icons.lock_outline, color: kMainColor),
+        suffixIcon: IconButton(
+          onPressed: () {
+            setState(() {
+              showPassword = !showPassword;
+            });
+          },
+          icon: Icon(
+            showPassword ? FeatherIcons.eyeOff : FeatherIcons.eye,
+            color: Colors.grey[600],
+          ),
+        ),
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: Colors.grey[300]!),
+        ),
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: Colors.grey[300]!),
+        ),
+        focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: kMainColor, width: 2),
+        ),
+        filled: true,
+        fillColor: Colors.grey[50],
+        contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
+      ),
+      validator: (value) {
+        if (value == null || value.isEmpty) {
+          return lang.S.of(context).passwordCannotBeEmpty;
+        } else if (value.length < 6) {
+          return lang.S.of(context).pleaseEnterABiggerPassword;
+        }
+        return null;
+      },
+    );
+  }
+
+  Widget _buildTermsSection(TextTheme textTheme) {
+    return Row(
+      children: [
+        Container(
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.circular(6),
+            border: Border.all(
+              color: kMainColor,
+              width: 2,
+            ),
+          ),
+          child: Checkbox(
+            materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
+            checkColor: Colors.white,
+            activeColor: kMainColor,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(6.0),
+            ),
+            value: true,
+            onChanged: (newValue) {
+              // TODO: Добавить логику для принятия условий
+            },
+          ),
+        ),
+        const SizedBox(width: 12),
+        Expanded(
+          child: RichText(
+            text: TextSpan(
+              text: 'Я принимаю ',
+              style: textTheme.bodyMedium?.copyWith(
+                color: Colors.grey[700],
+                fontSize: 14,
+              ),
+              children: [
+                TextSpan(
+                  text: 'Условия использования',
+                  style: TextStyle(
+                    color: kMainColor,
+                    fontWeight: FontWeight.w600,
+                    decoration: TextDecoration.underline,
+                  ),
+                ),
+                TextSpan(text: ' и '),
+                TextSpan(
+                  text: 'Политику конфиденциальности',
+                  style: TextStyle(
+                    color: kMainColor,
+                    fontWeight: FontWeight.w600,
+                    decoration: TextDecoration.underline,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildActionSection(TextTheme textTheme) {
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 5),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          SizedBox(
+            width: double.infinity,
+            height: 56,
+            child: ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                backgroundColor: kMainColor,
+                foregroundColor: Colors.white,
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(12),
+                ),
+                elevation: 0,
+                shadowColor: kMainColor.withOpacity(0.3),
+              ),
+              onPressed: () async {
+                if (isClicked) {
+                  return;
+                }
+                if (key.currentState?.validate() ?? false) {
+                  isClicked = true;
+                  EasyLoading.show();
+                  // TODO: Реализовать регистрацию
+                  await Future.delayed(const Duration(seconds: 2));
+                  EasyLoading.showSuccess('Регистрация успешна!');
+                  isClicked = false;
+                }
+              },
+              child: Text(
+                'Создать аккаунт',
+                style: const TextStyle(
+                  fontSize: 18,
+                  fontWeight: FontWeight.w600,
+                ),
+              ),
+            ),
+          ),
+          const SizedBox(height: 20),
+          Row(
+            children: [
+              Expanded(child: Divider(color: Colors.grey[300])),
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                child: Text(
+                  'или',
+                  style: TextStyle(
+                    color: Colors.grey[600],
+                    fontSize: 14,
+                  ),
+                ),
+              ),
+              Expanded(child: Divider(color: Colors.grey[300])),
+            ],
+          ),
+          const SizedBox(height: 20),
+          _buildSocialSignUpButtons(),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildSocialSignUpButtons() {
+    return Row(
+      children: [
+        Expanded(
+          child: _buildSocialButton(
+            icon: Icons.g_mobiledata,
+            label: 'Google',
+            color: Colors.red,
+            onTap: () => _handleGoogleSignUp(),
+          ),
+        ),
+        const SizedBox(width: 16),
+        Expanded(
+          child: _buildSocialButton(
+            icon: Icons.phone,
+            label: 'Телефон',
+            color: Colors.green,
+            onTap: () => _handlePhoneSignUp(),
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildSocialButton({
+    required IconData icon,
+    required String label,
+    required Color color,
+    required VoidCallback onTap,
+  }) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        height: 48,
+        decoration: BoxDecoration(
+          color: color.withOpacity(0.1),
+          borderRadius: BorderRadius.circular(12),
+          border: Border.all(color: color.withOpacity(0.3)),
+        ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(icon, color: color, size: 20),
+            const SizedBox(width: 8),
+            Text(
+              label,
+              style: TextStyle(
+                color: color,
+                fontSize: 14,
+                fontWeight: FontWeight.w600,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildSignInSection(TextTheme textTheme) {
+    return Container(
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 5),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          Text(
+            'Уже есть аккаунт?',
+            style: textTheme.bodyMedium?.copyWith(
+              color: Colors.grey[600],
+              fontSize: 16,
+            ),
+          ),
+          const SizedBox(height: 16),
+          SizedBox(
+            width: double.infinity,
+            height: 48,
+            child: OutlinedButton(
+              style: OutlinedButton.styleFrom(
+                foregroundColor: kMainColor,
+                side: BorderSide(color: kMainColor, width: 2),
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(12),
+                ),
+              ),
+              onPressed: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => const SignIn(),
+                  ),
+                );
+              },
+              child: Text(
+                'Войти',
+                style: TextStyle(
+                  fontSize: 16,
+                  fontWeight: FontWeight.w600,
+                  color: kMainColor,
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _handleGoogleSignUp() {
+    // TODO: Реализовать регистрацию через Google
+    EasyLoading.showInfo('Регистрация через Google будет добавлена');
+  }
+
+  void _handlePhoneSignUp() {
+    // TODO: Реализовать регистрацию через телефон
+    EasyLoading.showInfo('Регистрация через телефон будет добавлена');
   }
 }
