@@ -12,86 +12,438 @@ class CheckEMail extends StatefulWidget {
   _CheckEMailState createState() => _CheckEMailState();
 }
 
-class _CheckEMailState extends State<CheckEMail> {
+class _CheckEMailState extends State<CheckEMail> with TickerProviderStateMixin {
+  // Animation controllers
+  late AnimationController _fadeController;
+  late AnimationController _slideController;
+  late AnimationController _scaleController;
+  late AnimationController _pulseController;
+  late AnimationController _bounceController;
+
+  // Animations
+  late Animation<double> _fadeAnimation;
+  late Animation<Offset> _slideAnimation;
+  late Animation<double> _scaleAnimation;
+  late Animation<double> _pulseAnimation;
+  late Animation<double> _bounceAnimation;
+
+  @override
+  void initState() {
+    super.initState();
+    _initializeAnimations();
+  }
+
+  void _initializeAnimations() {
+    _fadeController = AnimationController(
+      duration: const Duration(milliseconds: 1000),
+      vsync: this,
+    );
+    _slideController = AnimationController(
+      duration: const Duration(milliseconds: 800),
+      vsync: this,
+    );
+    _scaleController = AnimationController(
+      duration: const Duration(milliseconds: 600),
+      vsync: this,
+    );
+    _pulseController = AnimationController(
+      duration: const Duration(milliseconds: 2000),
+      vsync: this,
+    );
+    _bounceController = AnimationController(
+      duration: const Duration(milliseconds: 1200),
+      vsync: this,
+    );
+
+    _fadeAnimation = Tween<double>(
+      begin: 0.0,
+      end: 1.0,
+    ).animate(CurvedAnimation(
+      parent: _fadeController,
+      curve: Curves.easeInOut,
+    ));
+
+    _slideAnimation = Tween<Offset>(
+      begin: const Offset(0, 0.5),
+      end: Offset.zero,
+    ).animate(CurvedAnimation(
+      parent: _slideController,
+      curve: Curves.easeOutCubic,
+    ));
+
+    _scaleAnimation = Tween<double>(
+      begin: 0.5,
+      end: 1.0,
+    ).animate(CurvedAnimation(
+      parent: _scaleController,
+      curve: Curves.elasticOut,
+    ));
+
+    _pulseAnimation = Tween<double>(
+      begin: 1.0,
+      end: 1.05,
+    ).animate(CurvedAnimation(
+      parent: _pulseController,
+      curve: Curves.easeInOut,
+    ));
+
+    _bounceAnimation = Tween<double>(
+      begin: 0.0,
+      end: 1.0,
+    ).animate(CurvedAnimation(
+      parent: _bounceController,
+      curve: Curves.elasticOut,
+    ));
+
+    // Start animations with delays
+    _fadeController.forward();
+    Future.delayed(const Duration(milliseconds: 200), () {
+      _slideController.forward();
+    });
+    Future.delayed(const Duration(milliseconds: 400), () {
+      _scaleController.forward();
+    });
+    Future.delayed(const Duration(milliseconds: 600), () {
+      _bounceController.forward();
+    });
+    _pulseController.repeat(reverse: true);
+  }
+
+  @override
+  void dispose() {
+    _fadeController.dispose();
+    _slideController.dispose();
+    _scaleController.dispose();
+    _pulseController.dispose();
+    _bounceController.dispose();
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
     return SafeArea(
       child: Scaffold(
-        body: Padding(
-          padding: const EdgeInsets.all(10.0),
-          child: Column(
-            children: [
-              Expanded(
-                flex: 5,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    const SizedBox(
-                      height: 100.0,
-                      width: 100.0,
-                      child: Image(
-                        image: AssetImage('images/mailbox.png'),
-                      ),
+        backgroundColor: kWhite,
+        body: CustomScrollView(
+          slivers: [
+            _buildSliverAppBar(),
+            SliverToBoxAdapter(
+              child: FadeTransition(
+                opacity: _fadeAnimation,
+                child: SlideTransition(
+                  position: _slideAnimation,
+                  child: Container(
+                    padding: const EdgeInsets.all(24),
+                    child: Column(
+                      children: [
+                        const SizedBox(height: 40),
+                        _buildEmailIcon(),
+                        const SizedBox(height: 40),
+                        _buildTitleSection(),
+                        const SizedBox(height: 30),
+                        _buildDescriptionSection(),
+                        const SizedBox(height: 40),
+                        _buildActionSection(),
+                        const SizedBox(height: 40),
+                      ],
                     ),
-                    Text(
-                      lang.S.of(context).gotEmail,
-                      style: GoogleFonts.poppins(
-                        fontSize: 25.0,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.black,
-                      ),
-                    ),
-                    Container(
-                      padding: const EdgeInsets.all(20.0),
-                      width: MediaQuery.of(context).size.width,
-                      child: Text(
-                        lang.S.of(context).sendEmail,
-                        textAlign: TextAlign.center,
-                        style: GoogleFonts.poppins(
-                          fontSize: 20.0,
-                          color: Colors.black,
-                        ),
-                      ),
-                    ),
-                    Text(
-                      'example@johndoe.com',
-                      style: GoogleFonts.poppins(
-                        fontSize: 20.0,
-                        color: Colors.black,
-                      ),
-                    ),
-                  ],
+                  ),
                 ),
               ),
-              Expanded(
-                flex: 1,
-                child: Column(
-                  children: [
-                    ButtonGlobalWithoutIcon(
-                      buttontext: lang.S.of(context).checkEmail,
-                      buttonDecoration: kButtonDecoration.copyWith(color: kMainColor),
-                      onPressed: null,
-                      buttonTextColor: Colors.white,
-                    ),
-                    TextButton(
-                      onPressed: () {
-                        Navigator.pushNamed(context, '/otp');
-                      },
-                      child: Text(
-                        lang.S.of(context).checkEmail,
-                        style: GoogleFonts.poppins(
-                          color: kMainColor,
-                        ),
-                      ),
-                    ),
-                  ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildSliverAppBar() {
+    return SliverAppBar(
+      expandedHeight: 120,
+      pinned: true,
+      backgroundColor: kMainColor,
+      flexibleSpace: FlexibleSpaceBar(
+        background: Container(
+          decoration: BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topLeft,
+              end: Alignment.bottomRight,
+              colors: [
+                kMainColor,
+                kMainColor.withOpacity(0.8),
+                kMainColor.withOpacity(0.6),
+              ],
+            ),
+          ),
+          child: Center(
+            child: ScaleTransition(
+              scale: _pulseAnimation,
+              child: Container(
+                padding: const EdgeInsets.all(16),
+                decoration: BoxDecoration(
+                  color: Colors.white.withOpacity(0.2),
+                  borderRadius: BorderRadius.circular(20),
                 ),
-              )
+                child: const Icon(
+                  Icons.email,
+                  size: 40,
+                  color: Colors.white,
+                ),
+              ),
+            ),
+          ),
+        ),
+        title: const Text(
+          'Check Email',
+          style: TextStyle(
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        centerTitle: true,
+      ),
+      elevation: 0,
+    );
+  }
+
+  Widget _buildEmailIcon() {
+    return ScaleTransition(
+      scale: _scaleAnimation,
+      child: Container(
+        padding: const EdgeInsets.all(32),
+        decoration: BoxDecoration(
+          color: kMainColor.withOpacity(0.1),
+          borderRadius: BorderRadius.circular(100),
+          border: Border.all(
+            color: kMainColor.withOpacity(0.3),
+            width: 2,
+          ),
+        ),
+        child: Container(
+          padding: const EdgeInsets.all(24),
+          decoration: BoxDecoration(
+            color: kMainColor,
+            borderRadius: BorderRadius.circular(76),
+            boxShadow: [
+              BoxShadow(
+                color: kMainColor.withOpacity(0.3),
+                spreadRadius: 8,
+                blurRadius: 20,
+                offset: const Offset(0, 8),
+              ),
             ],
+          ),
+          child: const Icon(
+            Icons.email_outlined,
+            size: 60,
+            color: Colors.white,
           ),
         ),
       ),
+    );
+  }
+
+  Widget _buildTitleSection() {
+    return Container(
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.withOpacity(0.1),
+            spreadRadius: 1,
+            blurRadius: 20,
+            offset: const Offset(0, 4),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          Text(
+            lang.S.of(context).gotEmail,
+            style: GoogleFonts.poppins(
+              fontSize: 28.0,
+              fontWeight: FontWeight.bold,
+              color: kMainColor,
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 16),
+          Container(
+            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+            decoration: BoxDecoration(
+              color: Colors.grey[50],
+              borderRadius: BorderRadius.circular(20),
+              border: Border.all(color: Colors.grey[200]!),
+            ),
+            child: Row(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Icon(
+                  Icons.info_outline,
+                  color: Colors.blue[600],
+                  size: 18,
+                ),
+                const SizedBox(width: 8),
+                Text(
+                  'Check your inbox',
+                  style: TextStyle(
+                    color: Colors.blue[600],
+                    fontWeight: FontWeight.w600,
+                    fontSize: 14,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildDescriptionSection() {
+    return Container(
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.grey[50],
+        borderRadius: BorderRadius.circular(20),
+        border: Border.all(color: Colors.grey[200]!),
+      ),
+      child: Column(
+        children: [
+          Icon(
+            Icons.mark_email_unread_outlined,
+            size: 32,
+            color: Colors.orange[600],
+          ),
+          const SizedBox(height: 16),
+          Text(
+            lang.S.of(context).sendEmail,
+            style: GoogleFonts.poppins(
+              fontSize: 18.0,
+              fontWeight: FontWeight.w600,
+              color: Colors.grey[800],
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 12),
+          Container(
+            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+            decoration: BoxDecoration(
+              color: Colors.white,
+              borderRadius: BorderRadius.circular(12),
+              border: Border.all(color: kMainColor.withOpacity(0.3)),
+            ),
+            child: Row(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Icon(
+                  Icons.email,
+                  color: kMainColor,
+                  size: 20,
+                ),
+                const SizedBox(width: 12),
+                Text(
+                  'example@johndoe.com',
+                  style: GoogleFonts.poppins(
+                    fontSize: 16.0,
+                    fontWeight: FontWeight.w600,
+                    color: kMainColor,
+                  ),
+                ),
+              ],
+            ),
+          ),
+          const SizedBox(height: 16),
+          Text(
+            "We've sent a verification link to your email address. Please check your inbox and click the link to verify your account.",
+            style: GoogleFonts.poppins(
+              color: Colors.grey[600],
+              fontSize: 14.0,
+              height: 1.5,
+            ),
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildActionSection() {
+    return Column(
+      children: [
+        BounceTransition(
+          animation: _bounceAnimation,
+          child: SizedBox(
+            width: double.infinity,
+            child: ButtonGlobalWithoutIcon(
+              buttontext: lang.S.of(context).checkEmail,
+              buttonDecoration: kButtonDecoration.copyWith(
+                color: kMainColor,
+                borderRadius: BorderRadius.circular(16),
+                boxShadow: [
+                  BoxShadow(
+                    color: kMainColor.withOpacity(0.3),
+                    spreadRadius: 2,
+                    blurRadius: 12,
+                    offset: const Offset(0, 4),
+                  ),
+                ],
+              ),
+              onPressed: () {
+                // Open email app or navigate to email
+                // This could open the default email app
+              },
+              buttonTextColor: Colors.white,
+            ),
+          ),
+        ),
+        const SizedBox(height: 16),
+        TextButton.icon(
+          onPressed: () {
+            Navigator.pushNamed(context, '/otp');
+          },
+          icon: Icon(
+            Icons.phone,
+            color: kMainColor,
+            size: 18,
+          ),
+          label: Text(
+            'Use Phone Number Instead',
+            style: GoogleFonts.poppins(
+              color: kMainColor,
+              fontWeight: FontWeight.w600,
+            ),
+          ),
+        ),
+        const SizedBox(height: 16),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              "Didn't receive the email? ",
+              style: GoogleFonts.poppins(
+                color: Colors.grey[600],
+                fontSize: 14,
+              ),
+            ),
+            TextButton(
+              onPressed: () {
+                // Resend email functionality
+              },
+              child: Text(
+                'Resend',
+                style: GoogleFonts.poppins(
+                  color: kMainColor,
+                  fontWeight: FontWeight.w600,
+                  fontSize: 14,
+                ),
+              ),
+            ),
+          ],
+        ),
+      ],
     );
   }
 }
