@@ -16,13 +16,69 @@ class ForgotPassword extends StatefulWidget {
   State<ForgotPassword> createState() => _ForgotPasswordState();
 }
 
-class _ForgotPasswordState extends State<ForgotPassword> {
+class _ForgotPasswordState extends State<ForgotPassword> with TickerProviderStateMixin {
   final _formKey = GlobalKey<FormState>();
   bool isClicked = false;
   final TextEditingController _emailController = TextEditingController();
 
+  // Анимации
+  late AnimationController _fadeController;
+  late AnimationController _slideController;
+  late AnimationController _scaleController;
+  late AnimationController _pulseController;
+  late Animation<double> _fadeAnimation;
+  late Animation<Offset> _slideAnimation;
+  late Animation<double> _scaleAnimation;
+  late Animation<double> _pulseAnimation;
+
+  @override
+  void initState() {
+    super.initState();
+    
+    // Инициализация анимаций
+    _fadeController = AnimationController(
+      duration: const Duration(milliseconds: 1000),
+      vsync: this,
+    );
+    _slideController = AnimationController(
+      duration: const Duration(milliseconds: 800),
+      vsync: this,
+    );
+    _scaleController = AnimationController(
+      duration: const Duration(milliseconds: 600),
+      vsync: this,
+    );
+    _pulseController = AnimationController(
+      duration: const Duration(milliseconds: 2000),
+      vsync: this,
+    );
+
+    _fadeAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
+      CurvedAnimation(parent: _fadeController, curve: Curves.easeInOut),
+    );
+    _slideAnimation = Tween<Offset>(
+      begin: const Offset(0, 0.5),
+      end: Offset.zero,
+    ).animate(CurvedAnimation(parent: _slideController, curve: Curves.easeOut));
+    _scaleAnimation = Tween<double>(begin: 0.8, end: 1.0).animate(
+      CurvedAnimation(parent: _scaleController, curve: Curves.elasticOut),
+    );
+    _pulseAnimation = Tween<double>(begin: 1.0, end: 1.05).animate(
+      CurvedAnimation(parent: _pulseController, curve: Curves.easeInOut),
+    );
+
+    _fadeController.forward();
+    _slideController.forward();
+    _scaleController.forward();
+    _pulseController.repeat(reverse: true);
+  }
+
   @override
   void dispose() {
+    _fadeController.dispose();
+    _slideController.dispose();
+    _scaleController.dispose();
+    _pulseController.dispose();
     _emailController.dispose();
     super.dispose();
   }
@@ -31,60 +87,300 @@ class _ForgotPasswordState extends State<ForgotPassword> {
   Widget build(BuildContext context) {
     TextTheme textTheme = Theme.of(context).textTheme;
     return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        titleSpacing: 16,
-        backgroundColor: kWhite,
-        surfaceTintColor: kWhite,
-        centerTitle: true,
-        title: Text(
-           // 'Forgot Password',
-          lang.S.of(context).forgotPassword,
-          style: textTheme.titleMedium?.copyWith(fontSize: 18),
-        ),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.fromLTRB(16.0, 20.0, 16.0, 0.0),
+      backgroundColor: const Color(0xFFF8F9FA),
+      body: CustomScrollView(
+        slivers: [
+          _buildSliverAppBar(textTheme),
+          SliverToBoxAdapter(
+            child: FadeTransition(
+              opacity: _fadeAnimation,
+              child: SlideTransition(
+                position: _slideAnimation,
+                child: Padding(
+                  padding: const EdgeInsets.fromLTRB(24.0, 20.0, 24.0, 0.0),
         child: Form(
           key: _formKey,
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        const SizedBox(height: 32),
+                        _buildLogoSection(),
+                        const SizedBox(height: 32),
+                        _buildWelcomeSection(textTheme),
+                        const SizedBox(height: 32),
+                        _buildFormSection(textTheme),
+                        const SizedBox(height: 32),
+                        _buildActionSection(textTheme),
+                        const SizedBox(height: 32),
+                        _buildHelpSection(textTheme),
+                        const SizedBox(height: 40),
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildSliverAppBar(TextTheme textTheme) {
+    return SliverAppBar(
+      expandedHeight: 120,
+      floating: false,
+      pinned: true,
+      backgroundColor: kMainColor,
+      elevation: 0,
+      automaticallyImplyLeading: false,
+      flexibleSpace: FlexibleSpaceBar(
+        title: Text(
+          lang.S.of(context).forgotPassword,
+          style: textTheme.titleSmall?.copyWith(
+            fontSize: 22,
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        background: Container(
+          decoration: BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topLeft,
+              end: Alignment.bottomRight,
+              colors: [
+                kMainColor,
+                kMainColor.withOpacity(0.8),
+                kMainColor.withOpacity(0.6),
+              ],
+            ),
+          ),
+          child: Center(
+            child: ScaleTransition(
+              scale: _pulseAnimation,
+              child: Icon(
+                Icons.lock_reset,
+                size: 60,
+                color: Colors.white.withOpacity(0.3),
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildLogoSection() {
+    return ScaleTransition(
+      scale: _scaleAnimation,
+      child: Container(
+        padding: const EdgeInsets.all(24),
+        decoration: BoxDecoration(
+          color: Colors.white,
+          borderRadius: BorderRadius.circular(20),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withOpacity(0.1),
+              blurRadius: 20,
+              offset: const Offset(0, 10),
+            ),
+          ],
+        ),
+        child: Column(
+          children: [
+            Container(
+              width: 80,
+              height: 80,
+              decoration: BoxDecoration(
+                color: kMainColor.withOpacity(0.1),
+                borderRadius: BorderRadius.circular(20),
+              ),
+              child: Icon(
+                Icons.lock_reset,
+                size: 40,
+                color: kMainColor,
+              ),
+            ),
+            const SizedBox(height: 16),
+            Text(
+              'Восстановление',
+              style: TextStyle(
+                fontSize: 24,
+                fontWeight: FontWeight.bold,
+                color: kMainColor,
+              ),
+            ),
+            Text(
+              'пароля',
+              style: TextStyle(
+                fontSize: 18,
+                fontWeight: FontWeight.w500,
+                color: Colors.grey[600],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildWelcomeSection(TextTheme textTheme) {
+    return Container(
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 5),
+          ),
+        ],
+      ),
+      child: Column(
             children: [
               Text(
-                // 'Forgot Password',
                 lang.S.of(context).forgotPassword,
-                style: textTheme.titleMedium?.copyWith(fontSize: 24.0),
-              ),
-              const SizedBox(height: 8.0),
+            style: textTheme.titleMedium?.copyWith(
+              fontSize: 28.0,
+              fontWeight: FontWeight.bold,
+              color: Colors.grey[800],
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 12),
               Text(
-                //'Reset password by using your email or phone number',
                 lang.S.of(context).reset,
-                style: textTheme.bodyMedium?.copyWith(color: kGreyTextColor,fontSize: 16),textAlign: TextAlign.center,
-              ),
-              const SizedBox(height: 24.0),
+            style: textTheme.bodyMedium?.copyWith(
+              color: Colors.grey[600],
+              fontSize: 16,
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 16),
+          Container(
+            padding: const EdgeInsets.all(16),
+            decoration: BoxDecoration(
+              color: Colors.blue.withOpacity(0.1),
+              borderRadius: BorderRadius.circular(12),
+              border: Border.all(color: Colors.blue.withOpacity(0.3)),
+            ),
+            child: Row(
+              children: [
+                Icon(Icons.info_outline, color: Colors.blue, size: 20),
+                const SizedBox(width: 12),
+                Expanded(
+                  child: Text(
+                    'Мы отправим инструкции по восстановлению пароля на ваш email',
+                    style: TextStyle(
+                      color: Colors.blue[700],
+                      fontSize: 14,
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildFormSection(TextTheme textTheme) {
+    return Container(
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 5),
+          ),
+        ],
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            'Email адрес',
+            style: textTheme.bodyMedium?.copyWith(
+              fontWeight: FontWeight.w600,
+              fontSize: 16,
+              color: Colors.grey[800],
+            ),
+          ),
+          const SizedBox(height: 12),
               TextFormField(
                 controller: _emailController,
                 keyboardType: TextInputType.emailAddress,
-                decoration:  kInputDecoration.copyWith(
-                  // labelText: 'Email',
+            decoration: InputDecoration(
                   labelText: lang.S.of(context).lableEmail,
-                  // hintText: 'Enter email address',
                   hintText: lang.S.of(context).hintEmail,
+              prefixIcon: Icon(Icons.email_outlined, color: kMainColor),
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(12),
+                borderSide: BorderSide(color: Colors.grey[300]!),
+              ),
+              enabledBorder: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(12),
+                borderSide: BorderSide(color: Colors.grey[300]!),
+              ),
+              focusedBorder: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(12),
+                borderSide: BorderSide(color: kMainColor, width: 2),
+              ),
+              filled: true,
+              fillColor: Colors.grey[50],
+              contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
                 ),
                 validator: (value) {
                   if (value == null || value.isEmpty) {
-                    //return 'Email can\'t be empty';
                     return lang.S.of(context).emailCannotBeEmpty;
                   } else if (!value.contains('@')) {
-                   // return 'Please enter a valid email';
                     return lang.S.of(context).pleaseEnterAValidEmail;
                   }
                   return null;
                 },
               ),
-              const SizedBox(height: 24.0),
-              UpdateButton(
-                onpressed: () async {
+        ],
+      ),
+    );
+  }
+
+  Widget _buildActionSection(TextTheme textTheme) {
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 5),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          SizedBox(
+            width: double.infinity,
+            height: 56,
+            child: ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                backgroundColor: kMainColor,
+                foregroundColor: Colors.white,
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(12),
+                ),
+                elevation: 0,
+                shadowColor: kMainColor.withOpacity(0.3),
+              ),
+              onPressed: () async {
                   if (isClicked) {
                     return;
                   }
@@ -109,221 +405,195 @@ class _ForgotPasswordState extends State<ForgotPassword> {
                     }
                   }
                 },
-                text:
+              child: Text(
                lang.S.of(context).continueE,
-                //'Continue',
+                style: const TextStyle(
+                  fontSize: 18,
+                  fontWeight: FontWeight.w600,
+                ),
+              ),
+            ),
+          ),
+          const SizedBox(height: 20),
+          Row(
+            children: [
+              Expanded(child: Divider(color: Colors.grey[300])),
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                child: Text(
+                  'или',
+                  style: TextStyle(
+                    color: Colors.grey[600],
+                    fontSize: 14,
+                  ),
+                ),
+              ),
+              Expanded(child: Divider(color: Colors.grey[300])),
+            ],
+          ),
+          const SizedBox(height: 20),
+          _buildAlternativeMethods(),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildAlternativeMethods() {
+    return Column(
+      children: [
+        _buildAlternativeButton(
+          icon: Icons.phone,
+          label: 'Восстановить через телефон',
+          color: Colors.green,
+          onTap: () => _handlePhoneRecovery(),
+        ),
+        const SizedBox(height: 12),
+        _buildAlternativeButton(
+          icon: Icons.security,
+          label: 'Ответить на секретный вопрос',
+          color: Colors.orange,
+          onTap: () => _handleSecurityQuestion(),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildAlternativeButton({
+    required IconData icon,
+    required String label,
+    required Color color,
+    required VoidCallback onTap,
+  }) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        width: double.infinity,
+        height: 48,
+        decoration: BoxDecoration(
+          color: color.withOpacity(0.1),
+          borderRadius: BorderRadius.circular(12),
+          border: Border.all(color: color.withOpacity(0.3)),
+        ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(icon, color: color, size: 20),
+            const SizedBox(width: 8),
+            Text(
+              label,
+              style: TextStyle(
+                color: color,
+                fontSize: 14,
+                fontWeight: FontWeight.w600,
+              ),
               ),
             ],
           ),
-        ),
       ),
-      // bottomNavigationBar: Padding(
-      //   padding: const EdgeInsets.fromLTRB(20.0, 0.0, 20.0, 20.0),
-      //   child: Column(
-      //     mainAxisSize: MainAxisSize.min,
-      //     crossAxisAlignment: CrossAxisAlignment.center,
-      //     mainAxisAlignment: MainAxisAlignment.center,
-      //     children: [
-      //       InkWell(
-      //         highlightColor: kMainColor.withOpacity(0.1),
-      //         borderRadius: BorderRadius.circular(3.0),
-      //         onTap: () => Navigator.pushReplacement(
-      //           context,
-      //           MaterialPageRoute(
-      //             builder: (context) => const SignUpScreen(),
-      //           ),
-      //         ),
-      //         hoverColor: kMainColor.withOpacity(0.1),
-      //         child: RichText(
-      //           text: TextSpan(
-      //             text: 'Don’t have an account? ',
-      //             style: textTheme.bodyMedium?.copyWith(color: kGreyTextColor),
-      //             children: [
-      //               TextSpan(
-      //                 text: 'Sign Up',
-      //                 style: textTheme.bodyMedium?.copyWith(color: kMainColor, fontWeight: FontWeight.bold),
-      //               )
-      //             ],
-      //           ),
-      //         ),
-      //       ),
-      //     ],
-      //   ),
-      // ),
     );
   }
-}
 
-// import 'dart:ui';
-//
-// import 'package:flutter/material.dart';
-// import 'package:influencer/widgets/theme/theme_constants.dart';
-// import '../../../../widgets/buttons widgets/button_widgets.dart';
-// import 'set_new_password.dart';
-//
-// class ForgotPassword extends StatefulWidget {
-//   const ForgotPassword({super.key});
-//
-//   @override
-//   State<ForgotPassword> createState() => _ForgotPasswordState();
-// }
-//
-// class _ForgotPasswordState extends State<ForgotPassword> {
-//   @override
-//   Widget build(BuildContext context) {
-//     TextTheme textTheme = Theme.of(context).textTheme;
-//     return Scaffold(
-//       backgroundColor: Colors.white,
-//       appBar: AppBar(
-//         automaticallyImplyLeading: false,
-//         titleSpacing: 16,
-//         leading: Padding(
-//           padding: const EdgeInsets.only(left: 16.0),
-//           child: Back(
-//             onPressed: () {
-//               Navigator.pop(context);
-//             },
-//           ),
-//         ),
-//         title: Text(
-//           'Forgot Password',
-//           style: textTheme.titleMedium,
-//         ),
-//       ),
-//       body: Padding(
-//         padding: const EdgeInsets.fromLTRB(16.0, 20.0, 16.0, 0.0),
-//         child: Column(
-//           crossAxisAlignment: CrossAxisAlignment.start,
-//           children: [
-//             Text(
-//               'Forgot Password',
-//               style: textTheme.titleMedium?.copyWith(fontSize: 30.0),
-//             ),
-//             const SizedBox(height: 8.0),
-//             Text(
-//               'Reset password by using your email or phone number',
-//               style: textTheme.bodyMedium?.copyWith(color: lightGreyTextColor),
-//             ),
-//             const SizedBox(height: 24.0),
-//             Text(
-//               'Email',
-//               style: textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.bold),
-//             ),
-//             const SizedBox(height: 8.0),
-//             TextFormField(
-//               keyboardType: TextInputType.emailAddress,
-//               decoration: const InputDecoration(
-//                 border: OutlineInputBorder(),
-//                 hintText: 'Enter email address',
-//               ),
-//               validator: (value) {
-//                 if (value == null || value.isEmpty) {
-//                   return 'Email can\'n be empty';
-//                 } else if (!value.contains('@')) {
-//                   return 'Please enter a valid email';
-//                 }
-//                 return null;
-//               },
-//               onSaved: (value) {},
-//             ),
-//             const SizedBox(height: 24.0),
-//             PrimaryButton(
-//               onPressed: () {
-//                 showDialog(
-//                   context: context,
-//                   builder: (BuildContext context) {
-//                     return BackdropFilter(
-//                       filter: ImageFilter.blur(sigmaX: 4, sigmaY: 4),
-//                       child: Dialog(
-//                         shape: RoundedRectangleBorder(
-//                           borderRadius: BorderRadius.circular(16.0),
-//                         ),
-//                         child: Padding(
-//                           padding: const EdgeInsets.fromLTRB(20.0, 38.0, 20.0, 28.0),
-//                           child: Column(
-//                             mainAxisSize: MainAxisSize.min,
-//                             crossAxisAlignment: CrossAxisAlignment.center,
-//                             mainAxisAlignment: MainAxisAlignment.center,
-//                             children: [
-//                               Text(
-//                                 'Verify Your Email',
-//                                 style: textTheme.titleMedium?.copyWith(fontSize: 24.0),
-//                               ),
-//                               const SizedBox(height: 16.0),
-//                               Text(
-//                                 'We have sent a confirmation email to',
-//                                 textAlign: TextAlign.center,
-//                                 style: textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.normal, color: lightGreyTextColor),
-//                               ),
-//                               Text(
-//                                 'sahidul11182@gmail.com',
-//                                 style: textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.bold),
-//                               ),
-//                               const SizedBox(height: 16.0),
-//                               Text(
-//                                 'It May be that the mail ended up in your spam folder.',
-//                                 textAlign: TextAlign.center,
-//                                 style: textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.normal, color: lightGreyTextColor),
-//                               ),
-//                               const SizedBox(height: 17.0),
-//                               PrimaryButton(
-//                                 onPressed: () {
-//                                   Navigator.pop(context);
-//                                   Navigator.push(
-//                                     context,
-//                                     MaterialPageRoute(
-//                                       builder: (context) => const ChangePassword(),
-//                                     ),
-//                                   );
-//                                 },
-//                                 text: 'Got It !',
-//                               ),
-//                             ],
-//                           ),
-//                         ),
-//                       ),
-//                     );
-//                   },
-//                 );
-//               },
-//               text: 'Continue',
-//             ),
-//           ],
-//         ),
-//       ),
-//       bottomNavigationBar: Padding(
-//         padding: const EdgeInsets.fromLTRB(20.0, 0.0, 20.0, 20.0),
-//         child: Column(
-//           mainAxisSize: MainAxisSize.min,
-//           crossAxisAlignment: CrossAxisAlignment.center,
-//           mainAxisAlignment: MainAxisAlignment.center,
-//           children: [
-//             InkWell(
-//               highlightColor: primaryColor.withOpacity(0.1),
-//               borderRadius: BorderRadius.circular(3.0),
-//               onTap: () => Navigator.push(
-//                 context,
-//                 MaterialPageRoute(
-//                   builder: (context) => const ForgotPassword(),
-//                 ),
-//               ),
-//               hoverColor: primaryColor.withOpacity(0.1),
-//               child: RichText(
-//                 text: TextSpan(
-//                   text: 'Don’t have an account? ',
-//                   style: textTheme.bodyMedium?.copyWith(color: lightGreyTextColor),
-//                   children: [
-//                     TextSpan(
-//                       text: 'Sign Up',
-//                       style: textTheme.bodyMedium?.copyWith(color: primaryColor, fontWeight: FontWeight.bold),
-//                     )
-//                   ],
-//                 ),
-//               ),
-//             ),
-//           ],
-//         ),
-//       ),
-//     );
-//   }
-// }
+  Widget _buildHelpSection(TextTheme textTheme) {
+    return Container(
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 5),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          Text(
+            'Нужна помощь?',
+            style: textTheme.bodyMedium?.copyWith(
+              color: Colors.grey[600],
+              fontSize: 16,
+            ),
+          ),
+          const SizedBox(height: 16),
+          Row(
+            children: [
+              Expanded(
+                child: _buildHelpButton(
+                  icon: Icons.support_agent,
+                  label: 'Поддержка',
+                  color: Colors.blue,
+                  onTap: () => _handleSupport(),
+                ),
+              ),
+              const SizedBox(width: 16),
+              Expanded(
+                child: _buildHelpButton(
+                  icon: Icons.chat_bubble_outline,
+                  label: 'Чат',
+                  color: Colors.green,
+                  onTap: () => _handleChat(),
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildHelpButton({
+    required IconData icon,
+    required String label,
+    required Color color,
+    required VoidCallback onTap,
+  }) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        height: 48,
+        decoration: BoxDecoration(
+          color: color.withOpacity(0.1),
+          borderRadius: BorderRadius.circular(12),
+          border: Border.all(color: color.withOpacity(0.3)),
+        ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(icon, color: color, size: 20),
+            const SizedBox(width: 8),
+            Text(
+              label,
+              style: TextStyle(
+                color: color,
+                fontSize: 14,
+                fontWeight: FontWeight.w600,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  void _handlePhoneRecovery() {
+    // TODO: Реализовать восстановление через телефон
+    EasyLoading.showInfo('Восстановление через телефон будет добавлено');
+  }
+
+  void _handleSecurityQuestion() {
+    // TODO: Реализовать восстановление через секретный вопрос
+    EasyLoading.showInfo('Восстановление через секретный вопрос будет добавлено');
+  }
+
+  void _handleSupport() {
+    // TODO: Реализовать поддержку
+    EasyLoading.showInfo('Поддержка будет добавлена');
+  }
+
+  void _handleChat() {
+    // TODO: Реализовать чат
+    EasyLoading.showInfo('Чат будет добавлен');
+  }
+}
