@@ -7,7 +7,7 @@ import '../../../constant.dart';
 import 'package:mobile_pos/generated/l10n.dart' as lang;
 
 class SetNewPassword extends StatefulWidget {
-  const SetNewPassword({Key? key,  required this.email}) : super(key: key);
+  const SetNewPassword({Key? key, required this.email}) : super(key: key);
 
   final String email;
 
@@ -15,7 +15,7 @@ class SetNewPassword extends StatefulWidget {
   State<SetNewPassword> createState() => _SetNewPasswordState();
 }
 
-class _SetNewPasswordState extends State<SetNewPassword> {
+class _SetNewPasswordState extends State<SetNewPassword> with TickerProviderStateMixin {
   final _formKey = GlobalKey<FormState>();
   bool isClicked = false;
   final TextEditingController _passwordController = TextEditingController();
@@ -24,8 +24,86 @@ class _SetNewPasswordState extends State<SetNewPassword> {
   bool showPassword = true;
   bool showConfirmPassword = true;
 
+  // Animation controllers
+  late AnimationController _fadeController;
+  late AnimationController _slideController;
+  late AnimationController _scaleController;
+  late AnimationController _pulseController;
+
+  // Animations
+  late Animation<double> _fadeAnimation;
+  late Animation<Offset> _slideAnimation;
+  late Animation<double> _scaleAnimation;
+  late Animation<double> _pulseAnimation;
+
+  @override
+  void initState() {
+    super.initState();
+    _initializeAnimations();
+  }
+
+  void _initializeAnimations() {
+    _fadeController = AnimationController(
+      duration: const Duration(milliseconds: 800),
+      vsync: this,
+    );
+    _slideController = AnimationController(
+      duration: const Duration(milliseconds: 600),
+      vsync: this,
+    );
+    _scaleController = AnimationController(
+      duration: const Duration(milliseconds: 500),
+      vsync: this,
+    );
+    _pulseController = AnimationController(
+      duration: const Duration(milliseconds: 2000),
+      vsync: this,
+    );
+
+    _fadeAnimation = Tween<double>(
+      begin: 0.0,
+      end: 1.0,
+    ).animate(CurvedAnimation(
+      parent: _fadeController,
+      curve: Curves.easeInOut,
+    ));
+
+    _slideAnimation = Tween<Offset>(
+      begin: const Offset(0, 0.3),
+      end: Offset.zero,
+    ).animate(CurvedAnimation(
+      parent: _slideController,
+      curve: Curves.easeOutCubic,
+    ));
+
+    _scaleAnimation = Tween<double>(
+      begin: 0.8,
+      end: 1.0,
+    ).animate(CurvedAnimation(
+      parent: _scaleController,
+      curve: Curves.elasticOut,
+    ));
+
+    _pulseAnimation = Tween<double>(
+      begin: 1.0,
+      end: 1.05,
+    ).animate(CurvedAnimation(
+      parent: _pulseController,
+      curve: Curves.easeInOut,
+    ));
+
+    _fadeController.forward();
+    _slideController.forward();
+    _scaleController.forward();
+    _pulseController.repeat(reverse: true);
+  }
+
   @override
   void dispose() {
+    _fadeController.dispose();
+    _slideController.dispose();
+    _scaleController.dispose();
+    _pulseController.dispose();
     _passwordController.dispose();
     _confirmPasswordController.dispose();
     super.dispose();
@@ -36,126 +114,381 @@ class _SetNewPasswordState extends State<SetNewPassword> {
     TextTheme textTheme = Theme.of(context).textTheme;
     return Scaffold(
       backgroundColor: Colors.white,
-      appBar: AppBar(
-        surfaceTintColor: kWhite,
-        backgroundColor: kWhite,
-        centerTitle: true,
-        titleSpacing: 16,
+      body: CustomScrollView(
+        slivers: [
+          _buildSliverAppBar(textTheme),
+          SliverToBoxAdapter(
+            child: FadeTransition(
+              opacity: _fadeAnimation,
+              child: SlideTransition(
+                position: _slideAnimation,
+                child: Container(
+                  padding: const EdgeInsets.all(24),
+                  child: Column(
+                    children: [
+                      _buildLogoSection(),
+                      const SizedBox(height: 30),
+                      _buildWelcomeSection(textTheme),
+                      const SizedBox(height: 30),
+                      _buildFormSection(textTheme),
+                      const SizedBox(height: 30),
+                      _buildActionSection(textTheme),
+                    ],
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildSliverAppBar(TextTheme textTheme) {
+    return SliverAppBar(
+      expandedHeight: 120,
+      pinned: true,
+      backgroundColor: kMainColor,
+      flexibleSpace: FlexibleSpaceBar(
+        background: Container(
+          decoration: BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topLeft,
+              end: Alignment.bottomRight,
+              colors: [
+                kMainColor,
+                kMainColor.withOpacity(0.8),
+                kMainColor.withOpacity(0.6),
+              ],
+            ),
+          ),
+          child: Center(
+            child: ScaleTransition(
+              scale: _pulseAnimation,
+              child: Container(
+                padding: const EdgeInsets.all(16),
+                decoration: BoxDecoration(
+                  color: Colors.white.withOpacity(0.2),
+                  borderRadius: BorderRadius.circular(20),
+                ),
+                child: const Icon(
+                  Icons.lock_reset,
+                  size: 40,
+                  color: Colors.white,
+                ),
+              ),
+            ),
+          ),
+        ),
         title: Text(
           lang.S.of(context).createNewPassword,
-          //'Create New Password',
-          style: textTheme.titleMedium?.copyWith(fontSize: 18),
+          style: const TextStyle(
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        centerTitle: true,
+      ),
+      elevation: 0,
+    );
+  }
+
+  Widget _buildLogoSection() {
+    return ScaleTransition(
+      scale: _scaleAnimation,
+      child: Container(
+        padding: const EdgeInsets.all(24),
+        decoration: BoxDecoration(
+          color: kMainColor.withOpacity(0.1),
+          borderRadius: BorderRadius.circular(20),
+          border: Border.all(
+            color: kMainColor.withOpacity(0.3),
+            width: 1,
+          ),
+        ),
+        child: Column(
+          children: [
+            Container(
+              padding: const EdgeInsets.all(16),
+              decoration: BoxDecoration(
+                color: kMainColor,
+                borderRadius: BorderRadius.circular(16),
+              ),
+              child: const Icon(
+                Icons.lock_reset,
+                size: 32,
+                color: Colors.white,
+              ),
+            ),
+            const SizedBox(height: 16),
+            const Text(
+              'POSPro',
+              style: TextStyle(
+                fontSize: 24,
+                fontWeight: FontWeight.bold,
+                color: kMainColor,
+              ),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              'Password Reset',
+              style: TextStyle(
+                fontSize: 14,
+                color: Colors.grey[600],
+              ),
+            ),
+          ],
         ),
       ),
-      body: SingleChildScrollView(
-        child: Padding(
-          padding: const EdgeInsets.fromLTRB(16.0, 20.0, 16.0, 0.0),
-          child: Form(
-            key: _formKey,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.center,
+    );
+  }
+
+  Widget _buildWelcomeSection(TextTheme textTheme) {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(16),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.withOpacity(0.1),
+            spreadRadius: 1,
+            blurRadius: 10,
+            offset: const Offset(0, 2),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          Text(
+            lang.S.of(context).setUpNewPassword,
+            style: textTheme.titleMedium?.copyWith(
+              fontSize: 24.0,
+              fontWeight: FontWeight.bold,
+              color: kMainColor,
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 12),
+          Text(
+            lang.S.of(context).resetPassword,
+            style: textTheme.bodyMedium?.copyWith(
+              color: kGreyTextColor,
+              fontSize: 16,
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 16),
+          Container(
+            padding: const EdgeInsets.all(12),
+            decoration: BoxDecoration(
+              color: kMainColor.withOpacity(0.1),
+              borderRadius: BorderRadius.circular(8),
+              border: Border.all(color: kMainColor.withOpacity(0.3)),
+            ),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.center,
               children: [
+                Icon(
+                  Icons.email,
+                  color: kMainColor,
+                  size: 18,
+                ),
+                const SizedBox(width: 8),
                 Text(
-                 lang.S.of(context).setUpNewPassword,
-                 // 'Set Up New Password',
-                  style: textTheme.titleMedium?.copyWith(fontSize: 24.0),
-                ),
-                const SizedBox(height: 8.0),
-                Text(
-                  lang.S.of(context).resetPassword,
-                  //'Reset your password to recovery and log in your account',
-                  style: textTheme.bodyMedium?.copyWith(color: kGreyTextColor,fontSize: 16),textAlign: TextAlign.center,
-                ),
-                const SizedBox(height: 24.0),
-                TextFormField(
-                  controller: _passwordController,
-                  keyboardType: TextInputType.text,
-                  obscureText: showPassword,
-                  decoration: kInputDecoration.copyWith(
-                    // border: const OutlineInputBorder(),
-                    hintText: '********',
-                    //labelText: 'New Password',
-                    labelText: lang.S.of(context).newPassword,
-                    suffixIcon: IconButton(
-                      onPressed: () {
-                        setState(() {
-                          showPassword = !showPassword;
-                        });
-                      },
-                      icon: Icon(
-                        showPassword ? FeatherIcons.eyeOff : FeatherIcons.eye,
-                        color: kGreyTextColor,
-                      ),
-                    ),
+                  widget.email,
+                  style: TextStyle(
+                    color: kMainColor,
+                    fontWeight: FontWeight.w600,
                   ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      //return 'Password can\'t be empty';
-                      return lang.S.of(context).passwordCannotBeEmpty;
-                    } else if (value.length < 6) {
-                      //return 'Please enter a bigger password';
-                      return lang.S.of(context).pleaseEnterABiggerPassword;
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 20.0),
-                TextFormField(
-                  controller: _confirmPasswordController,
-                  keyboardType: TextInputType.text,
-                  obscureText: showConfirmPassword,
-                  decoration: kInputDecoration.copyWith(
-                    border: const OutlineInputBorder(),
-                    ///labelText: 'Confirm Password',
-                    labelText: lang.S.of(context).confirmPassword,
-                    hintText: '********',
-                    suffixIcon: IconButton(
-                      onPressed: () {
-                        setState(() {
-                          showConfirmPassword = !showConfirmPassword;
-                        });
-                      },
-                      icon: Icon(
-                        showConfirmPassword ? FeatherIcons.eyeOff : FeatherIcons.eye,
-                        color: kGreyTextColor,
-                      ),
-                    ),
-                  ),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      //return 'Password can\'t be empty';
-                      return lang.S.of(context).passwordCannotBeEmpty;
-                    } else if (value != _passwordController.text) {
-                      //return 'Passwords do not match';
-                      return lang.S.of(context).passwordsDoNotMatch;
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 24.0),
-                UpdateButton(
-                  onpressed: () async {
-                    if (isClicked) {
-                      return;
-                    }
-                    if (_formKey.currentState?.validate() ?? false) {
-                      isClicked = true;
-                      EasyLoading.show();
-                      ForgotPassRepo repo = ForgotPassRepo();
-                      if (await repo.resetPass(email: widget.email, password: _confirmPasswordController.text, context: context)) {
-                        Navigator.pop(context);
-                      } else {
-                        isClicked = false;
-                      }
-                    }
-                  },
-                  text: lang.S.of(context).save,
-                  //'Save',
                 ),
               ],
             ),
           ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildFormSection(TextTheme textTheme) {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(16),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.withOpacity(0.1),
+            spreadRadius: 1,
+            blurRadius: 10,
+            offset: const Offset(0, 2),
+          ),
+        ],
+      ),
+      child: Form(
+        key: _formKey,
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            const Text(
+              'New Password',
+              style: TextStyle(
+                fontSize: 16,
+                fontWeight: FontWeight.w600,
+                color: kMainColor,
+              ),
+            ),
+            const SizedBox(height: 12),
+            _buildPasswordField(
+              controller: _passwordController,
+              labelText: lang.S.of(context).newPassword,
+              hintText: 'Enter new password',
+              showPassword: showPassword,
+              onTogglePassword: () {
+                setState(() {
+                  showPassword = !showPassword;
+                });
+              },
+              validator: (value) {
+                if (value == null || value.isEmpty) {
+                  return lang.S.of(context).passwordCannotBeEmpty;
+                } else if (value.length < 6) {
+                  return lang.S.of(context).pleaseEnterABiggerPassword;
+                }
+                return null;
+              },
+            ),
+            const SizedBox(height: 20),
+            const Text(
+              'Confirm Password',
+              style: TextStyle(
+                fontSize: 16,
+                fontWeight: FontWeight.w600,
+                color: kMainColor,
+              ),
+            ),
+            const SizedBox(height: 12),
+            _buildPasswordField(
+              controller: _confirmPasswordController,
+              labelText: lang.S.of(context).confirmPassword,
+              hintText: 'Confirm new password',
+              showPassword: showConfirmPassword,
+              onTogglePassword: () {
+                setState(() {
+                  showConfirmPassword = !showConfirmPassword;
+                });
+              },
+              validator: (value) {
+                if (value == null || value.isEmpty) {
+                  return lang.S.of(context).passwordCannotBeEmpty;
+                } else if (value != _passwordController.text) {
+                  return lang.S.of(context).passwordsDoNotMatch;
+                }
+                return null;
+              },
+            ),
+          ],
         ),
+      ),
+    );
+  }
+
+  Widget _buildPasswordField({
+    required TextEditingController controller,
+    required String labelText,
+    required String hintText,
+    required bool showPassword,
+    required VoidCallback onTogglePassword,
+    required String? Function(String?) validator,
+  }) {
+    return TextFormField(
+      controller: controller,
+      keyboardType: TextInputType.text,
+      obscureText: showPassword,
+      decoration: InputDecoration(
+        labelText: labelText,
+        hintText: hintText,
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: Colors.grey[300]!),
+        ),
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: Colors.grey[300]!),
+        ),
+        focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: const BorderSide(color: kMainColor, width: 2),
+        ),
+        filled: true,
+        fillColor: Colors.grey[50],
+        contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+        suffixIcon: IconButton(
+          onPressed: onTogglePassword,
+          icon: Icon(
+            showPassword ? FeatherIcons.eyeOff : FeatherIcons.eye,
+            color: kGreyTextColor,
+          ),
+        ),
+      ),
+      validator: validator,
+    );
+  }
+
+  Widget _buildActionSection(TextTheme textTheme) {
+    return SizedBox(
+      width: double.infinity,
+      height: 50,
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(
+          backgroundColor: kMainColor,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(12),
+          ),
+          elevation: 2,
+        ),
+        onPressed: isClicked
+            ? null
+            : () async {
+                if (isClicked) {
+                  return;
+                }
+                if (_formKey.currentState?.validate() ?? false) {
+                  setState(() {
+                    isClicked = true;
+                  });
+                  EasyLoading.show();
+                  ForgotPassRepo repo = ForgotPassRepo();
+                  if (await repo.resetPass(
+                    email: widget.email,
+                    password: _confirmPasswordController.text,
+                    context: context,
+                  )) {
+                    Navigator.pop(context);
+                  } else {
+                    setState(() {
+                      isClicked = false;
+                    });
+                  }
+                }
+              },
+        child: isClicked
+            ? const SizedBox(
+                width: 20,
+                height: 20,
+                child: CircularProgressIndicator(
+                  strokeWidth: 2,
+                  valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+                ),
+              )
+            : Text(
+                lang.S.of(context).save,
+                style: const TextStyle(
+                  color: Colors.white,
+                  fontSize: 16,
+                  fontWeight: FontWeight.w600,
+                ),
+              ),
       ),
     );
   }
