@@ -1,272 +1,761 @@
 // ignore: import_of_legacy_library_into_null_safe
 
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:mobile_pos/GlobalComponents/button_global.dart';
-import 'package:mobile_pos/constant.dart';
-import 'package:mobile_pos/generated/l10n.dart' as lang;
-
-import '../../currency.dart';
+import 'package:provider/provider.dart';
+import '../../Provider/app_provider.dart';
+import '../../Provider/auth_provider.dart';
+import '../../Provider/analytics_provider.dart';
 
 class ProfileScreen extends StatefulWidget {
   const ProfileScreen({Key? key}) : super(key: key);
 
   @override
-  // ignore: library_private_types_in_public_api
   _ProfileScreenState createState() => _ProfileScreenState();
 }
 
-class _ProfileScreenState extends State<ProfileScreen> {
+class _ProfileScreenState extends State<ProfileScreen>
+    with TickerProviderStateMixin {
+  late AnimationController _fadeController;
+  late AnimationController _slideController;
+  late AnimationController _scaleController;
+  late Animation<double> _fadeAnimation;
+  late Animation<Offset> _slideAnimation;
+  late Animation<double> _scaleAnimation;
+
+  @override
+  void initState() {
+    super.initState();
+    _fadeController = AnimationController(
+      duration: const Duration(milliseconds: 800),
+      vsync: this,
+    );
+    _slideController = AnimationController(
+      duration: const Duration(milliseconds: 600),
+      vsync: this,
+    );
+    _scaleController = AnimationController(
+      duration: const Duration(milliseconds: 400),
+      vsync: this,
+    );
+
+    _fadeAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
+      CurvedAnimation(parent: _fadeController, curve: Curves.easeInOut),
+    );
+    _slideAnimation = Tween<Offset>(
+      begin: const Offset(0, 0.3),
+      end: Offset.zero,
+    ).animate(CurvedAnimation(parent: _slideController, curve: Curves.easeOut));
+    _scaleAnimation = Tween<double>(begin: 0.8, end: 1.0).animate(
+      CurvedAnimation(parent: _scaleController, curve: Curves.elasticOut),
+    );
+
+    _fadeController.forward();
+    _slideController.forward();
+    _scaleController.forward();
+  }
+
+  @override
+  void dispose() {
+    _fadeController.dispose();
+    _slideController.dispose();
+    _scaleController.dispose();
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: kWhite,
-      appBar: AppBar(
-        iconTheme: const IconThemeData(color: Colors.black),
-        title: Text(
-          lang.S.of(context).contactDetials,
-          style: GoogleFonts.poppins(
-            color: Colors.black,
-          ),
-        ),
-        centerTitle: true,
-        backgroundColor: Colors.white,
-        elevation: 4.0,
-        actions: const [
-          Icon(
-            Icons.mode_edit_outline,
-            color: kGreyTextColor,
-          ),
-          Padding(
-            padding: EdgeInsets.all(8.0),
-            child: Icon(
-              Icons.delete_outline,
-              color: kGreyTextColor,
+      backgroundColor: const Color(0xFFF8F9FA),
+      body: CustomScrollView(
+        slivers: [
+          _buildSliverAppBar(),
+          SliverToBoxAdapter(
+            child: FadeTransition(
+              opacity: _fadeAnimation,
+              child: SlideTransition(
+                position: _slideAnimation,
+                child: Column(
+                  children: [
+                    _buildProfileHeader(),
+                    _buildQuickStats(),
+                    _buildProfileSections(),
+                    _buildSettingsSection(),
+                    const SizedBox(height: 30),
+                  ],
+                ),
+              ),
             ),
           ),
         ],
       ),
-      body: Center(
+    );
+  }
+
+  Widget _buildSliverAppBar() {
+    return SliverAppBar(
+      expandedHeight: 120,
+      floating: false,
+      pinned: true,
+      backgroundColor: Theme.of(context).primaryColor,
+      elevation: 0,
+      flexibleSpace: FlexibleSpaceBar(
+        title: const Text(
+          'Профиль',
+          style: TextStyle(
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        background: Container(
+          decoration: BoxDecoration(
+            gradient: LinearGradient(
+              begin: Alignment.topLeft,
+              end: Alignment.bottomRight,
+              colors: [
+                Theme.of(context).primaryColor,
+                Theme.of(context).primaryColor.withOpacity(0.8),
+              ],
+            ),
+          ),
+        ),
+      ),
+      actions: [
+        IconButton(
+          icon: const Icon(Icons.edit, color: Colors.white),
+          onPressed: () => _showEditProfileDialog(),
+        ),
+        IconButton(
+          icon: const Icon(Icons.settings, color: Colors.white),
+          onPressed: () => _showSettingsDialog(),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildProfileHeader() {
+    return Container(
+      margin: const EdgeInsets.all(20),
+      padding: const EdgeInsets.all(24),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(20),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.1),
+            blurRadius: 20,
+            offset: const Offset(0, 10),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          ScaleTransition(
+            scale: _scaleAnimation,
+            child: Container(
+              width: 120,
+              height: 120,
+              decoration: BoxDecoration(
+                shape: BoxShape.circle,
+                gradient: LinearGradient(
+                  colors: [
+                    Theme.of(context).primaryColor,
+                    Theme.of(context).primaryColor.withOpacity(0.7),
+                  ],
+                ),
+                boxShadow: [
+                  BoxShadow(
+                    color: Theme.of(context).primaryColor.withOpacity(0.3),
+                    blurRadius: 20,
+                    offset: const Offset(0, 10),
+                  ),
+                ],
+              ),
+              child: const Icon(
+                Icons.person,
+                size: 60,
+                color: Colors.white,
+              ),
+            ),
+          ),
+          const SizedBox(height: 20),
+          Text(
+            'Александр Петров',
+            style: TextStyle(
+              fontSize: 28,
+              fontWeight: FontWeight.bold,
+              color: Colors.grey[800],
+            ),
+          ),
+          const SizedBox(height: 8),
+          Text(
+            'Менеджер по продажам',
+            style: TextStyle(
+              fontSize: 16,
+              color: Colors.grey[600],
+            ),
+          ),
+          const SizedBox(height: 8),
+          Text(
+            'alex.petrov@company.com',
+            style: TextStyle(
+              fontSize: 14,
+              color: Colors.grey[500],
+            ),
+          ),
+          const SizedBox(height: 20),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+            children: [
+              _buildActionButton(
+                icon: Icons.phone,
+                label: 'Позвонить',
+                color: Colors.green,
+                onTap: () => _showCallDialog(),
+              ),
+              _buildActionButton(
+                icon: Icons.message,
+                label: 'Сообщение',
+                color: Colors.blue,
+                onTap: () => _showMessageDialog(),
+              ),
+              _buildActionButton(
+                icon: Icons.email,
+                label: 'Email',
+                color: Colors.orange,
+                onTap: () => _showEmailDialog(),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildActionButton({
+    required IconData icon,
+    required String label,
+    required Color color,
+    required VoidCallback onTap,
+  }) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+        decoration: BoxDecoration(
+          color: color.withOpacity(0.1),
+          borderRadius: BorderRadius.circular(12),
+          border: Border.all(color: color.withOpacity(0.3)),
+        ),
         child: Column(
           children: [
-            const SizedBox(
-              height: 20.0,
+            Icon(icon, color: color, size: 24),
+            const SizedBox(height: 4),
+            Text(
+              label,
+              style: TextStyle(
+                color: color,
+                fontSize: 12,
+                fontWeight: FontWeight.w500,
+              ),
             ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildQuickStats() {
+    return Container(
+      margin: const EdgeInsets.symmetric(horizontal: 20),
+      child: Row(
+        children: [
+          Expanded(
+            child: _buildStatCard(
+              title: 'Продажи',
+              value: '₽ 45,230',
+              icon: Icons.trending_up,
+              color: Colors.green,
+              onTap: () => _showSalesDetails(),
+            ),
+          ),
+          const SizedBox(width: 12),
+          Expanded(
+            child: _buildStatCard(
+              title: 'Заказы',
+              value: '127',
+              icon: Icons.shopping_cart,
+              color: Colors.blue,
+              onTap: () => _showOrdersDetails(),
+            ),
+          ),
+          const SizedBox(width: 12),
+          Expanded(
+            child: _buildStatCard(
+              title: 'Клиенты',
+              value: '89',
+              icon: Icons.people,
+              color: Colors.purple,
+              onTap: () => _showCustomersDetails(),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildStatCard({
+    required String title,
+    required String value,
+    required IconData icon,
+    required Color color,
+    required VoidCallback onTap,
+  }) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        padding: const EdgeInsets.all(16),
+        decoration: BoxDecoration(
+          color: Colors.white,
+          borderRadius: BorderRadius.circular(16),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withOpacity(0.05),
+              blurRadius: 10,
+              offset: const Offset(0, 5),
+            ),
+          ],
+        ),
+        child: Column(
+          children: [
             Container(
-              height: 100.0,
-              width: 100.0,
+              padding: const EdgeInsets.all(8),
               decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(10.0),
+                color: color.withOpacity(0.1),
+                borderRadius: BorderRadius.circular(12),
               ),
-              child: const Image(
-                image: AssetImage('images/profileimage.png'),
-                height: 100.0,
-                width: 100.0,
-              ),
+              child: Icon(icon, color: color, size: 24),
             ),
-            const SizedBox(
-              height: 10.0,
-            ),
+            const SizedBox(height: 12),
             Text(
-              lang.S.of(context).clarence,
-              style: GoogleFonts.poppins(
+              value,
+              style: TextStyle(
+                fontSize: 20,
                 fontWeight: FontWeight.bold,
-                fontSize: 25.0,
-                color: Colors.black,
+                color: Colors.grey[800],
               ),
             ),
+            const SizedBox(height: 4),
             Text(
-              '+1 (732) 343-4343',
-              style: GoogleFonts.poppins(
-                fontSize: 20.0,
-                color: Colors.black,
+              title,
+              style: TextStyle(
+                fontSize: 12,
+                color: Colors.grey[600],
+              ),
+              textAlign: TextAlign.center,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildProfileSections() {
+    return Container(
+      margin: const EdgeInsets.all(20),
+      child: Column(
+        children: [
+          _buildSectionCard(
+            title: 'Личная информация',
+            icon: Icons.person_outline,
+            children: [
+              _buildInfoRow('Имя', 'Александр'),
+              _buildInfoRow('Фамилия', 'Петров'),
+              _buildInfoRow('Телефон', '+7 (999) 123-45-67'),
+              _buildInfoRow('Email', 'alex.petrov@company.com'),
+              _buildInfoRow('Должность', 'Менеджер по продажам'),
+              _buildInfoRow('Отдел', 'Продажи'),
+            ],
+          ),
+          const SizedBox(height: 16),
+          _buildSectionCard(
+            title: 'Рабочая информация',
+            icon: Icons.work_outline,
+            children: [
+              _buildInfoRow('ID сотрудника', 'EMP-001'),
+              _buildInfoRow('Дата приема', '15.03.2022'),
+              _buildInfoRow('Стаж', '2 года 3 месяца'),
+              _buildInfoRow('Руководитель', 'Иван Сидоров'),
+              _buildInfoRow('Офис', 'Москва, ул. Тверская, 1'),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildSectionCard({
+    required String title,
+    required IconData icon,
+    required List<Widget> children,
+  }) {
+    return Container(
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(16),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 5),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          Container(
+            padding: const EdgeInsets.all(20),
+            decoration: BoxDecoration(
+              color: Theme.of(context).primaryColor.withOpacity(0.1),
+              borderRadius: const BorderRadius.only(
+                topLeft: Radius.circular(16),
+                topRight: Radius.circular(16),
               ),
             ),
-            const SizedBox(
-              height: 10.0,
-            ),
-            Row(
+            child: Row(
               children: [
-                Expanded(
-                  child: Padding(
-                    padding: const EdgeInsets.all(10.0),
-                    child: IconWithText(
-                      bgColor: const Color(0xFFF1F7F7),
-                      title: lang.S.of(context).call,
-                      iconData: Icons.call_outlined,
-                      iconColor: Colors.black,
-                    ),
-                  ),
+                Icon(
+                  icon,
+                  color: Theme.of(context).primaryColor,
+                  size: 24,
                 ),
-                Expanded(
-                  child: Padding(
-                    padding: EdgeInsets.all(10.0),
-                    child: IconWithText(
-                      bgColor: kMainColor,
-                      title: lang.S.of(context).messege,
-                      iconData: Icons.message,
-                      iconColor: Colors.white,
-                    ),
-                  ),
-                ),
-                 Expanded(
-                  child: Padding(
-                    padding: EdgeInsets.all(10.0),
-                    child: IconWithText(
-                      bgColor: Color(0xFFF1F7F7),
-                      title:lang.S.of(context).email,
-                      //'Email',
-                      iconData: Icons.email_outlined,
-                      iconColor: Colors.black,
-                    ),
+                const SizedBox(width: 12),
+                Text(
+                  title,
+                  style: TextStyle(
+                    fontSize: 18,
+                    fontWeight: FontWeight.bold,
+                    color: Colors.grey[800],
                   ),
                 ),
               ],
             ),
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Card(
-                child: Padding(
-                  padding: const EdgeInsets.all(20.0),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        lang.S.of(context).dailyTransaction,
-                        style: GoogleFonts.poppins(
-                          fontSize: 20.0,
-                          color: Colors.black,
-                        ),
-                      ),
-                      const SizedBox(
-                        height: 10.0,
-                      ),
-                      Transactions(
-                        title: lang.S.of(context).sales,
-                        amount: '$currency 1509.09',
-                        tranColor: Colors.black,
-                        pressed: () {},
-                      ),
-                      Transactions(
-                        title: lang.S.of(context).due,
-                        amount: '$currency 509.09',
-                        tranColor: kGreyTextColor,
-                        pressed: null,
-                      ),
-                      Transactions(
-                        title: lang.S.of(context).promo,
-                        amount: '$currency 109.09',
-                        tranColor: kGreyTextColor,
-                        pressed: null,
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-            const SizedBox(
-              height: 20.0,
-            ),
-            ButtonGlobalWithoutIcon(
-                buttontext: lang.S.of(context).send, buttonDecoration: kButtonDecoration.copyWith(color: kMainColor), onPressed: null, buttonTextColor: Colors.white),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-// ignore: must_be_immutable
-class Transactions extends StatelessWidget {
-  Transactions({
-    Key? key,
-    required this.title,
-    required this.amount,
-    required this.tranColor,
-    required this.pressed,
-  }) : super(key: key);
-  final String title;
-  final String amount;
-  final Color tranColor;
-  // ignore: prefer_typing_uninitialized_variables
-  var pressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: pressed,
-      child: Row(
-        children: [
-          Text(
-            title,
-            style: GoogleFonts.poppins(
-              fontSize: 20.0,
-              color: tranColor,
-            ),
           ),
-          const Spacer(),
-          Text(
-            amount,
-            style: GoogleFonts.poppins(
-              fontSize: 20.0,
-              color: tranColor,
-            ),
-          ),
-          const SizedBox(
-            width: 10.0,
-          ),
-          Icon(
-            Icons.arrow_forward_ios,
-            size: 15.0,
-            color: tranColor,
+          Padding(
+            padding: const EdgeInsets.all(20),
+            child: Column(children: children),
           ),
         ],
       ),
     );
   }
-}
 
-class IconWithText extends StatelessWidget {
-  const IconWithText({
-    Key? key,
-    required this.iconData,
-    required this.title,
-    required this.bgColor,
-    required this.iconColor,
-  }) : super(key: key);
-  final String title;
-  final IconData iconData;
-  final Color bgColor;
-  final Color iconColor;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 100.0,
-      width: 100.0,
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(10.0),
-        color: bgColor,
-      ),
-      child: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              iconData,
-              size: 32,
-              color: iconColor,
-            ),
-            Text(
-              title,
-              maxLines: 1,
-              style: GoogleFonts.poppins(
-                fontSize: 20.0,
-                color: iconColor,
+  Widget _buildInfoRow(String label, String value) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 8),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          SizedBox(
+            width: 120,
+            child: Text(
+              label,
+              style: TextStyle(
+                fontSize: 14,
+                color: Colors.grey[600],
+                fontWeight: FontWeight.w500,
               ),
+            ),
+          ),
+          Expanded(
+            child: Text(
+              value,
+              style: TextStyle(
+                fontSize: 14,
+                color: Colors.grey[800],
+                fontWeight: FontWeight.w600,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildSettingsSection() {
+    return Container(
+      margin: const EdgeInsets.symmetric(horizontal: 20),
+      child: Column(
+        children: [
+          _buildSettingsCard(
+            title: 'Настройки',
+            icon: Icons.settings,
+            onTap: () => _showSettingsDialog(),
+          ),
+          const SizedBox(height: 12),
+          _buildSettingsCard(
+            title: 'Безопасность',
+            icon: Icons.security,
+            onTap: () => _showSecurityDialog(),
+          ),
+          const SizedBox(height: 12),
+          _buildSettingsCard(
+            title: 'Уведомления',
+            icon: Icons.notifications,
+            onTap: () => _showNotificationsDialog(),
+          ),
+          const SizedBox(height: 12),
+          _buildSettingsCard(
+            title: 'Выйти',
+            icon: Icons.logout,
+            onTap: () => _showLogoutDialog(),
+            isDestructive: true,
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildSettingsCard({
+    required String title,
+    required IconData icon,
+    required VoidCallback onTap,
+    bool isDestructive = false,
+  }) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        padding: const EdgeInsets.all(16),
+        decoration: BoxDecoration(
+          color: Colors.white,
+          borderRadius: BorderRadius.circular(12),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withOpacity(0.05),
+              blurRadius: 10,
+              offset: const Offset(0, 5),
             ),
           ],
         ),
+        child: Row(
+          children: [
+            Container(
+              padding: const EdgeInsets.all(8),
+              decoration: BoxDecoration(
+                color: isDestructive
+                    ? Colors.red.withOpacity(0.1)
+                    : Theme.of(context).primaryColor.withOpacity(0.1),
+                borderRadius: BorderRadius.circular(8),
+              ),
+              child: Icon(
+                icon,
+                color: isDestructive
+                    ? Colors.red
+                    : Theme.of(context).primaryColor,
+                size: 20,
+              ),
+            ),
+            const SizedBox(width: 16),
+            Expanded(
+              child: Text(
+                title,
+                style: TextStyle(
+                  fontSize: 16,
+                  fontWeight: FontWeight.w500,
+                  color: isDestructive ? Colors.red : Colors.grey[800],
+                ),
+              ),
+            ),
+            Icon(
+              Icons.arrow_forward_ios,
+              size: 16,
+              color: Colors.grey[400],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  // Диалоги
+  void _showEditProfileDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Редактировать профиль'),
+        content: const Text('Функция редактирования профиля будет добавлена'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showSettingsDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Настройки'),
+        content: const Text('Настройки приложения будут добавлены'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showCallDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Звонок'),
+        content: const Text('Функция звонка будет добавлена'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showMessageDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Сообщение'),
+        content: const Text('Функция сообщений будет добавлена'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showEmailDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Email'),
+        content: const Text('Функция отправки email будет добавлена'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showSalesDetails() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Детали продаж'),
+        content: const Text('Подробная информация о продажах будет добавлена'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showOrdersDetails() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Детали заказов'),
+        content: const Text('Подробная информация о заказах будет добавлена'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showCustomersDetails() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Детали клиентов'),
+        content: const Text('Подробная информация о клиентах будет добавлена'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showSecurityDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Безопасность'),
+        content: const Text('Настройки безопасности будут добавлены'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showNotificationsDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Уведомления'),
+        content: const Text('Настройки уведомлений будут добавлены'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _showLogoutDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Выйти'),
+        content: const Text('Вы уверены, что хотите выйти из приложения?'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('Отмена'),
+          ),
+          TextButton(
+            onPressed: () {
+              Navigator.pop(context);
+              // Логика выхода будет добавлена
+            },
+            child: const Text('Выйти'),
+          ),
+        ],
       ),
     );
   }
